% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{Adjust_Const_Type}
\alias{Adjust_Const_Type}
\title{Adjust the constant label}
\usage{
Adjust_Const_Type(Const_Type_Full)
}
\arguments{
\item{Const_Type_Full}{Type of constraint (Full Name)}
}
\description{
Adjust the constant label
}
\keyword{internal}
