% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{adjust_delta}
\alias{adjust_delta}
\title{Adjust delta for numerical differentiation}
\usage{
adjust_delta(f, x, delta, i, fx0, direction)
}
\arguments{
\item{f}{function which contains vector (J x T) valued function handle}

\item{x}{parameter values}

\item{delta}{initial delta value}

\item{i}{index of the parameter being adjusted}

\item{fx0}{initial function value}

\item{direction}{direction of adjustment (1 for positive, -1 for negative)}
}
\value{
adjusted delta value
}
\description{
Adjust delta for numerical differentiation
}
\keyword{internal}
