% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{AdjustOptm_BS}
\alias{AdjustOptm_BS}
\title{Gathers the estimate of the bootstrap draws}
\usage{
AdjustOptm_BS(ModelType, ModelBootstrap, Draw_Opt, Economies, tt)
}
\arguments{
\item{ModelType}{String-vector containing the label of the model to be estimated}

\item{ModelBootstrap}{List to store the bootstrap set}

\item{Draw_Opt}{List of model estimated parameters}

\item{Economies}{String-vector containing the names of the economies which are part of the economic system}

\item{tt}{Number of the bootstrap draw}
}
\description{
Gathers the estimate of the bootstrap draws
}
\keyword{internal}
