% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{AdjustPathIRFs}
\alias{AdjustPathIRFs}
\title{Generate paths to save IRFs/GIRFs graphs}
\usage{
AdjustPathIRFs(OutputType, ElementType, PathsGraphs, Economies, ModelType)
}
\arguments{
\item{OutputType}{available options are "IRF" and "GIRF"}

\item{ElementType}{available options are "Factors" and "Yields"}

\item{PathsGraphs}{desired path to save the graphs}

\item{Economies}{Economies of the economic system}

\item{ModelType}{Desired estimated model type}
}
\description{
Generate paths to save IRFs/GIRFs graphs
}
\keyword{internal}
