% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{AdjustYieldsDates}
\alias{AdjustYieldsDates}
\title{Makes sure that the time series of yields and risk factors have coincident sample spans}
\usage{
AdjustYieldsDates(Yields, PdynamicsFactors, Economies)
}
\arguments{
\item{Yields}{time series of bond yields (CJ x T1 )}

\item{PdynamicsFactors}{time series of risk factors (F x T2)}

\item{Economies}{string-vector containing the names of the economies of the system.}
}
\description{
Makes sure that the time series of yields and risk factors have coincident sample spans
}
\keyword{internal}
