% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Methods_All.R
\name{autoplot.ATSMModelBoot}
\alias{autoplot.ATSMModelBoot}
\title{Autoplot method for ATSMModelBoot objects}
\usage{
\method{autoplot}{ATSMModelBoot}(x, NumOutPE, type, ...)
}
\arguments{
\item{x}{An object of class 'ATSMModelBoot'}

\item{NumOutPE}{An object of class 'ATSMNumOutputs': point estimates of the numerical outputs}

\item{type}{Plot type:  one of "IRF", "FEVD", "GIRF", "GFEVD" (each must be suffixed with "_Factors" or
"_Yields"). For JLL-based models, an additional "_Ortho" suffix produces
orthogonalized outputs. All inputs must end by "_Boot" as a reference to the bootstrap procedure.}

\item{...}{Additional arguments (not used)}
}
\description{
Autoplot method for ATSMModelBoot objects
}
