% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Methods_All.R
\name{autoplot.ATSMNumOutputs}
\alias{autoplot.ATSMNumOutputs}
\title{Autoplot method for ATSMNumOutputs objects}
\usage{
\method{autoplot}{ATSMNumOutputs}(x, type, ...)
}
\arguments{
\item{x}{An object of class 'ATSMNumOutputs'}

\item{type}{Plot type: "RiskFactors", "Fit", "TermPremia", or one of
"IRF", "FEVD", "GIRF", "GFEVD" (each must be suffixed with "_Factors" or
"_Yields"). For JLL-based models, an additional "_Ortho" suffix produces
orthogonalized outputs.}

\item{...}{Additional arguments (not used)}
}
\description{
Autoplot method for ATSMNumOutputs objects
}
