% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Methods_All.R
\name{autoplot}
\alias{autoplot}
\title{Autoplot generic function}
\usage{
autoplot(x, ...)
}
\arguments{
\item{x}{Object to plot}

\item{...}{Additional arguments passed to methods}
}
\description{
Autoplot generic function
}
