% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{Aux_BlockDiag}
\alias{Aux_BlockDiag}
\title{Transformation of the block diagonal parameters (auxiliary form)}
\usage{
Aux_BlockDiag(ParaValue, Const_Type, FactorLabels, Economies)
}
\arguments{
\item{ParaValue}{Constrained parameter value}

\item{Const_Type}{Type of constraint}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Transformation of the block diagonal parameters (auxiliary form)
}
\keyword{internal}
