% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{Aux_BoundDiag}
\alias{Aux_BoundDiag}
\title{Transformation of the bounded parameters (auxiliary form)}
\usage{
Aux_BoundDiag(ParaValue, Const_Type, lb, ub)
}
\arguments{
\item{ParaValue}{Constrained parameter value}

\item{Const_Type}{Type of constraint}

\item{lb}{lower bound}

\item{ub}{upper bound}
}
\description{
Transformation of the bounded parameters (auxiliary form)
}
\keyword{internal}
