% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{Aux_Jordan}
\alias{Aux_Jordan}
\title{Transformation of the Jordan-related parameters (auxiliary form)}
\usage{
Aux_Jordan(ParaValue, Const_Type, FactorLabels, Economies, JLLinputs)
}
\arguments{
\item{ParaValue}{Constrained parameter}

\item{Const_Type}{Type of constraint}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{JLLinputs}{list of necessary inputs for the estimation of JLL-based models (see "JLL" function)}
}
\description{
Transformation of the Jordan-related parameters (auxiliary form)
}
\keyword{internal}
