% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{Aux_jordan_OneCountry}
\alias{Aux_jordan_OneCountry}
\title{Auxiliary function for a single-country specification}
\usage{
Aux_jordan_OneCountry(ParaValue, Const_Type)
}
\arguments{
\item{ParaValue}{Constrained parameter}

\item{Const_Type}{Type of constraint}
}
\description{
Auxiliary function for a single-country specification
}
\references{
Le, A., & Singleton, K. J. (2018). Small Package of Matlab Routines for
Estimation of Some Term Structure Models. EABCN Training School.\cr
This function offers an independent R implementation that is informed
by the conceptual framework outlined in Le and Singleton (2018), but adapted to the
present modeling context.
}
\keyword{internal}
