% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{Aux_PSD}
\alias{Aux_PSD}
\title{Transformation of a PSD matrix (auxiliary form)}
\usage{
Aux_PSD(ParaValue, Const_Type)
}
\arguments{
\item{ParaValue}{Constrained parameter value}

\item{Const_Type}{Type of constraint}
}
\description{
Transformation of a PSD matrix (auxiliary form)
}
\keyword{internal}
