% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{Bias_Correc_VAR}
\alias{Bias_Correc_VAR}
\title{Estimates an unbiased VAR(1) using stochastic approximation (Bauer, Rudebusch and Wu, 2012)}
\usage{
Bias_Correc_VAR(
  ModelType,
  BRWinputs,
  RiskFactors,
  Economies,
  FactorLabels,
  GVARinputs = NULL,
  JLLinputs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ModelType}{A character vector indicating the model type to be estimated.}

\item{BRWinputs}{A list containing the necessary inputs for the BRW model estimation:
\enumerate{
\item \code{Cent_Measure}: Determines whether "Mean" or "Median"-unbiased estimation is desired.
\item \code{gamma}: Numeric. Adjustment parameter between 0 and 1. Default is 0.5.
\item \code{N_iter}: Integer. Number of iterations for the stochastic approximation algorithm after burn-in. Default is 5,000.
\item \code{N_burn}: Integer. Number of burn-in iterations. Default is 15% of \code{N_iter}.
\item \code{B}: Integer. Number of bootstrap samples per iteration for calculating the noisy measure of the biased estimator's mean or median. Default is 50.
\item \code{check}: Logical. Indicates whether to perform a closeness check. Default is TRUE.
\item \code{B_check}: Integer. Number of bootstrap samples for the closeness check. Default is 100,000.
\item \code{Eigen_rest}: Numeric. Restriction on the largest eigenvalue under the P-measure. Default is 1.
}}

\item{RiskFactors}{A numeric matrix (T x F) representing the time series of risk factors.}

\item{Economies}{A character vector containing the names of the economies included in the system.}

\item{FactorLabels}{A list of character vectors with labels for all variables in the model.}

\item{GVARinputs}{List. Inputs for GVAR model estimation (see \code{\link{GVAR}} function). Default is NULL.}

\item{JLLinputs}{List. Inputs for JLL model estimation (see \code{\link{JLL}} function). Default is NULL.}

\item{verbose}{verbose Logical flag controlling function messaging. Default is TRUE.}
}
\value{
Bias-corrected VAR parameters based on the framework of Bauer, Rudebusch and Wu (2012). The list contains:
\enumerate{
\item \code{KOZ_BC}: estimated intercept (F x 1);
\item \code{K1Z_BC}: estimated feedback matrix (F x F);
\item \code{SSZ_BC}: estimated variance-covariance matrix (F x F);
\item \code{dist}: root mean square distance (scalar);
}
}
\description{
Estimates an unbiased VAR(1) using stochastic approximation (Bauer, Rudebusch and Wu, 2012)
}
\examples{
\donttest{
data(CM_Factors)
Factors <- t(RiskFactors[1:7,])

BRWinputs <- list(Cent_Measure = "Mean", gamma = 0.4, N_iter = 1000, N_burn = 100,
                  B = 10, check = 1, B_check = 5000)

Economies <- "China"
N <- 3
ModelType <- "JPS original"
FactorLabels <- NULL

BRWpara <- Bias_Correc_VAR(ModelType, BRWinputs, Factors, Economies, FactorLabels)
}

}
\references{
Bauer, Rudebusch and, Wu (2012). "Correcting Estimation Bias in Dynamic Term Structure Models" \cr
An R implementation inspired by the methodology described therein. Related Matlab routines
are available on Cynthia Wu's website (https://sites.google.com/view/jingcynthiawu/).
}
