% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBoundsBootstrap.R
\name{Boot_DataGraphYield_perShock}
\alias{Boot_DataGraphYield_perShock}
\title{Generates the desired bootstrap graphs}
\usage{
Boot_DataGraphYield_perShock(
  NumOutBounds,
  NumOutPE,
  IdxShock,
  nmResponse,
  Lab_Int,
  ModelType,
  FacDim,
  YieldDim,
  Horiz,
  Economies = NULL,
  Ortho = FALSE
)
}
\arguments{
\item{NumOutBounds}{numerical output set from the bootstrap analysis}

\item{NumOutPE}{numerical output set from the point estimate analysis}

\item{IdxShock}{index associated with the shock variable}

\item{nmResponse}{Label of the response variable}

\item{Lab_Int}{Output types "IRF" or "FEVD"}

\item{ModelType}{desired model type}

\item{FacDim}{dimension from bond yield set}

\item{YieldDim}{dimension from the P-dynamics}

\item{Horiz}{horizon of analysis}

\item{Economies}{name of economies forming the economic system}

\item{Ortho}{Option for orthogonal outputs, for JLL models. Default is FALSE.}
}
\description{
Generates the desired bootstrap graphs
}
\keyword{internal}
