% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBoundsBootstrap.R
\name{Boot_graph_template}
\alias{Boot_graph_template}
\title{Builds template from bootstrap-related graphs}
\usage{
Boot_graph_template(ALLdata, nmResponse)
}
\arguments{
\item{ALLdata}{data-frame containing the necessary data for building the graphs}

\item{nmResponse}{string containing the name of the response variable}
}
\description{
Builds template from bootstrap-related graphs
}
\keyword{internal}
