% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBoundsBootstrap.R
\name{Boot_Yields_Graphs}
\alias{Boot_Yields_Graphs}
\title{Build P-dynamic graphs after the bootstrap implementation}
\usage{
Boot_Yields_Graphs(
  NumOutBounds,
  NumOutPE,
  ModelType,
  FacDim,
  YielDim,
  Horiz,
  Economies,
  PathsGraphs,
  OutInt,
  WishYieldGraphs,
  WishYieldGraphsOrtho = NULL
)
}
\arguments{
\item{NumOutBounds}{numerical output set from the bootstrap analysis}

\item{NumOutPE}{numerical output set from the point estimate analysis}

\item{ModelType}{desired model type}

\item{FacDim}{dimension of the risk factor set}

\item{Horiz}{horizon of numerical outputs}

\item{Economies}{Economies that are part of the economic system}

\item{PathsGraphs}{Path to save the desired graphs}

\item{OutInt}{Available option are "IRF" or "FEVD"}

\item{WishYieldGraphs}{Binary variable reflecting the graphs of interest}

\item{WishYieldGraphsOrtho}{Binary variable reflecting the graphs of interest (orthogonalized version). Default is NULL}
}
\description{
Build P-dynamic graphs after the bootstrap implementation
}
\keyword{internal}
