% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBoundsBootstrap.R
\name{BootstrapBoundsSet}
\alias{BootstrapBoundsSet}
\title{Builds the confidence bounds and graphs (Bootstrap set)}
\usage{
BootstrapBoundsSet(
  ModelType,
  ModelBootstrap,
  NumOutPE,
  InputsForOutputs,
  Economies
)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelBootstrap}{list containing the complete set of model parameters after the bootstrap estimation procedure}

\item{NumOutPE}{point estimate from the numerical outputs (see the outputs of the "NumOutputs" function)}

\item{InputsForOutputs}{list conataining the desired inputs for the construction of IRFs, GIRFs, FEVDs, and GFEVDs}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Builds the confidence bounds and graphs (Bootstrap set)
}
