% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinorOptimizationFunctions.R
\name{bound2x}
\alias{bound2x}
\title{Transform a number bounded between a lower bound and upper bound to x by:}
\usage{
bound2x(y, lb, ub)
}
\arguments{
\item{y}{Number to be transformed (scalar)}

\item{lb}{lower bound (scalar)}

\item{ub}{upper bound (scalar)}
}
\description{
Transform a number bounded between a lower bound and upper bound to x by:
}
\references{
This function is based on the "bound2x" function by Le and Singleton (2018). \cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
 Available at: https://cepr.org/40029
}
