% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{Build_xvec}
\alias{Build_xvec}
\title{Obtain the auxiliary values corresponding to each parameter, its size and its name}
\usage{
Build_xvec(ListInputSet, Economies, FactorLabels, JLLinputs = NULL)
}
\arguments{
\item{ListInputSet}{variable inputs used in the optimization (see "optimization" function)}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{FactorLabels}{list of necessary inputs for the estimation of JLL-based models (see \code{JLL} function)}

\item{JLLinputs}{Necessary inputs for the estimation of the JLL-based models}
}
\description{
Obtain the auxiliary values corresponding to each parameter, its size and its name
}
\references{
This function is a based on the "getx" function by Le and Singleton (2018). \cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
}
\keyword{internal}
