% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{BuildFEVDlist}
\alias{BuildFEVDlist}
\title{Build the list of IRF and GIRF for both factors and bond yields}
\usage{
BuildFEVDlist(
  NumOut,
  Economies,
  ModelType,
  FEVDhoriz,
  FacDim,
  YieldsDim,
  OutputType
)
}
\arguments{
\item{NumOut}{list of computed outputs containing the model fit, IRFs, FEVDs, GIRFs, GFEVDs and Term Premia}

\item{Economies}{Economies of the economic system}

\item{ModelType}{Desired model type}

\item{FEVDhoriz}{time-horizon of the FEVD and GFEVD}

\item{FacDim}{dimension of the risk factor vector}

\item{YieldsDim}{dimension of the model set of yields}

\item{OutputType}{available option are 'FEVD' and 'GFEVD'}
}
\description{
Build the list of IRF and GIRF for both factors and bond yields
}
\keyword{internal}
