% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{BuildGVAR}
\alias{BuildGVAR}
\title{Build the GVAR(1) from the country-specific VARX(1,1,1)}
\usage{
BuildGVAR(ParaVARX, GlobalPara, GVARinputs, DomLabels, GlobalLabels, N)
}
\arguments{
\item{ParaVARX}{Set of VARX model parameters}

\item{GlobalPara}{Set of marginal model parameters}

\item{GVARinputs}{List of inputs for GVAR-based models}

\item{DomLabels}{string-based vector containing label of the domestic risk factors}

\item{GlobalLabels}{string-based vector containing label of the global risk factors}

\item{N}{number of country-specific spanned factors (scalar)}
}
\description{
Build the GVAR(1) from the country-specific VARX(1,1,1)
}
\keyword{internal}
