% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{BuildIRFlist}
\alias{BuildIRFlist}
\title{Build the list of IRF and GIRF for both factors and bond yields}
\usage{
BuildIRFlist(NumOut, Economies, ModelType, IRFhoriz, FacDim, OutputType)
}
\arguments{
\item{NumOut}{list of computed outputs containing the model fit, IRFs, FEVDs, GIRFs, GFEVDs and Term Premia}

\item{Economies}{Economies of the economic system}

\item{ModelType}{Desired model type}

\item{IRFhoriz}{time-horizon of the IRF and GIRF}

\item{FacDim}{dimension of the risk factor vector}

\item{OutputType}{available option are 'IRF' and 'GIRF'}
}
\description{
Build the list of IRF and GIRF for both factors and bond yields
}
\keyword{internal}
