% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{BuildLinkMat}
\alias{BuildLinkMat}
\title{Build country-specific link matrices}
\usage{
BuildLinkMat(GVARinputs, N, M)
}
\arguments{
\item{GVARinputs}{List of inputs for GVAR-based models}

\item{N}{number of country-specific spanned factors (scalar)}

\item{M}{number of country-specific unspanned factors (scalar)}
}
\description{
Build country-specific link matrices
}
\keyword{internal}
