% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{BUnspannedAdapSep}
\alias{BUnspannedAdapSep}
\title{Transform B_spanned into B_unspanned for sepQ models}
\usage{
BUnspannedAdapSep(G, M, ModelPara_Short, Economies, Economy, ModelType)
}
\arguments{
\item{G}{number of global unspanned factors}

\item{M}{number of domestic unspanned factors per country}

\item{ModelPara_Short}{list of model parameter estimates (See the "Optimization" function)}

\item{Economies}{complet set of economies of the economic system}

\item{Economy}{specific economy under study}

\item{ModelType}{a string-vector containing the label of the model to be estimated}
}
\description{
Transform B_spanned into B_unspanned for sepQ models
}
\keyword{internal}
