% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputsBootstrap.R
\name{BUnspannedAdapSep_BS}
\alias{BUnspannedAdapSep_BS}
\title{Obtain the full form of B unspanned for "sep Q" models within the bootstrap setting}
\usage{
BUnspannedAdapSep_BS(G, M, ModelParaBoot, Economies, Economy, ModelType, tt)
}
\arguments{
\item{G}{number of global unspanned factors}

\item{M}{number of country-specific domestic unspanned factors}

\item{ModelParaBoot}{list of model parameter estimates (see the "Optimization" function) after a bootstrap draw}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{Economy}{string-vector containing the names of the economy under study}

\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{tt}{number of the bootstrap draw}
}
\description{
Obtain the full form of B unspanned for "sep Q" models within the bootstrap setting
}
\keyword{internal}
