% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{Check_comparison_NoBC}
\alias{Check_comparison_NoBC}
\title{check how close the mean or median of the bias-corrected aproach is from the non-corrected approach}
\usage{
Check_comparison_NoBC(
  K1Z_BC,
  K1Z_NoBC,
  B_check,
  RiskFactors,
  GVARinputs,
  JLLinputs,
  FactorLabels,
  Economies,
  ModelType,
  Use_Mean,
  verbose
)
}
\arguments{
\item{K1Z_BC}{feedback matrix after the bias correction procedure}

\item{K1Z_NoBC}{feedback matrix before the bias correction procedure}

\item{B_check}{number of bootstrap samples used in the closeness check}

\item{RiskFactors}{time series of the risk factors (F x T)}

\item{GVARinputs}{inputs used in the estimation of the GVAR-based models (see "GVAR" function). Default is set to NULL}

\item{JLLinputs}{inputs used in the estimation of the JLL-based models (see "JLL" function). Default is set to NULL}

\item{FactorLabels}{string-list based which contains the labels of all variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{Use_Mean}{Choose between the mean or the median across the bootstrap iterations.}

\item{verbose}{Logical flag controlling function messaging.}
}
\description{
check how close the mean or median of the bias-corrected aproach is from the non-corrected approach
}
\keyword{internal}
