% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{Check_label_consistency}
\alias{Check_label_consistency}
\title{Check consistency of labels (economies, domestic and global variables)}
\usage{
Check_label_consistency(
  Economies,
  DomesticMacroFac,
  GlobalMacroFac,
  DomVarLab,
  GloVarLab
)
}
\arguments{
\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{DomesticMacroFac}{time series of the country-specific domestic risk factors (C(M+N) x T)}

\item{GlobalMacroFac}{time series of the global risk factors (G X T)}

\item{DomVarLab}{string-vector containing the names of the desired domestic risk factors}

\item{GloVarLab}{string-vector containing the names of the desired global risk factors}
}
\description{
Check consistency of labels (economies, domestic and global variables)
}
\keyword{internal}
