% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{check_numeric}
\alias{check_numeric}
\title{Check for presence of NAs and infinite in numeric variables}
\usage{
check_numeric(x, name)
}
\arguments{
\item{x}{variable to be tested (numeric)}

\item{name}{variable labels}
}
\description{
Check for presence of NAs and infinite in numeric variables
}
\keyword{internal}
