% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{CheckInput_K1X}
\alias{CheckInput_K1X}
\title{Input validation for the 'FeedMat_Q' function}
\usage{
CheckInput_K1X(Yields, mat_vec, time_step)
}
\arguments{
\item{Yields}{matrix containing country-specific yields  (J x T)}

\item{mat_vec}{vector of maturities expressed in years (J x 1)}

\item{time_step}{time unit of the model (scalar).}
}
\description{
Input validation for the 'FeedMat_Q' function
}
\keyword{internal}
