% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{CheckInputsGVAR}
\alias{CheckInputsGVAR}
\title{Check consistency of the inputs provided in GVARinputs}
\usage{
CheckInputsGVAR(GVARinputs, N)
}
\arguments{
\item{GVARinputs}{List of inputs for GVAR-based models}

\item{N}{number of country-specific spanned factors (scalar)}
}
\description{
Check consistency of the inputs provided in GVARinputs
}
\keyword{internal}
