% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{CheckJLLinputs}
\alias{CheckJLLinputs}
\title{Check consistency of the inputs provided in JLL-based models}
\usage{
CheckJLLinputs(RiskFactorsNonOrtho, JLLinputs)
}
\arguments{
\item{RiskFactorsNonOrtho}{Risk factors before the orthogonalization (FxT)}

\item{JLLinputs}{List of necessary inputs to estimate JLL-based setups}
}
\description{
Check consistency of the inputs provided in JLL-based models
}
\keyword{internal}
