% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{CheckNumericalPrecision}
\alias{CheckNumericalPrecision}
\title{Check Numerical Precision Issues of K1_root matrix}
\usage{
CheckNumericalPrecision(matrix_in, tolerance = 1e-10)
}
\arguments{
\item{matrix_in}{K1_root matrix}

\item{tolerance}{Numerical tolerance}
}
\value{
List with precision indicators
}
\description{
Check Numerical Precision Issues of K1_root matrix
}
\keyword{internal}
