% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{ChecksOOS}
\alias{ChecksOOS}
\title{Preliminary checks for inputs provided for the performing out-of-sample forecasting}
\usage{
ChecksOOS(t0Forecast, t0Sample, nForecasts, ForecastType, TimeLength)
}
\arguments{
\item{t0Forecast}{Index of the last set of observations in the information set at the first forecasting round}

\item{t0Sample}{Index of the first set of observations in the information set at the first forecasting round}

\item{nForecasts}{Number of forecasting sets generated}

\item{ForecastType}{Forecast type. Available options are "Rolling" and "Expanding".}

\item{TimeLength}{Time-series dimension of the model}
}
\description{
Preliminary checks for inputs provided for the performing out-of-sample forecasting
}
\keyword{internal}
