% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{clean_labels}
\alias{clean_labels}
\title{Modify variable labels to make legends more readable}
\usage{
clean_labels(x)
}
\arguments{
\item{x}{variable labels prior to the modification}
}
\description{
Modify variable labels to make legends more readable
}
\keyword{internal}
