% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{CleanOrthoJLL_Boot}
\alias{CleanOrthoJLL_Boot}
\title{Clean unnecessary outputs of JLL models in the bootstrap setup}
\usage{
CleanOrthoJLL_Boot(ModelBootstrap, ndraws, ModelType)
}
\arguments{
\item{ModelBootstrap}{List of outputs to store bootstrap draws}

\item{ndraws}{Total number of bootstrap draws}

\item{ModelType}{A character vector indicating the model type to be estimated}
}
\description{
Clean unnecessary outputs of JLL models in the bootstrap setup
}
\keyword{internal}
