% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A0N__BnAn.R
\name{Compute_BnX_AnX}
\alias{Compute_BnX_AnX}
\title{Compute the latent loading AnX and BnX}
\usage{
Compute_BnX_AnX(mat, N, K1XQ, r0, dX, SSX, Economies, ModelType, Lab_SingleQ)
}
\arguments{
\item{mat}{vector of maturities (J x 1). Maturities are in multiples of the discrete interval used in the model}

\item{N}{number of country-specific spanned factors}

\item{K1XQ}{risk neutral feedback matrix (N x N)}

\item{r0}{the long run risk neutral mean of the short rate (scalar)}

\item{dX}{state loadings for the one-period rate (1xN). Default is a vector of ones}

\item{SSX}{the covariance matrix of the errors (N x N)}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{Lab_SingleQ}{string-vector containing the labels of the models estimated on a country-by-country basis}
}
\description{
Compute the latent loading AnX and BnX
}
\keyword{internal}
