% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{Compute_EP}
\alias{Compute_EP}
\title{Compute the expected component for all models}
\usage{
Compute_EP(
  ModelPara,
  ModelType,
  UnitYields,
  matAdjUnit,
  N,
  rhoList,
  Economies,
  FactorLabels,
  WishFP = FALSE,
  matMIN = FALSE,
  matMAX = FALSE
)
}
\arguments{
\item{ModelPara}{list of model parameter estimates}

\item{ModelType}{Desired model type}

\item{UnitYields}{Available options: "Month" and "Year"}

\item{matAdjUnit}{Adjusted vector of matutities}

\item{N}{number of country-specific spanned factors}

\item{rhoList}{List of risk-neutral parameters}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{FactorLabels}{List of factor labels}

\item{WishFP}{If users wants to compute the forward premia. Default is FALSE.}

\item{matMIN}{For the forward premia, the shortest maturity of the remium of interest}

\item{matMAX}{For the forward premia, the longest maturity of the remium of interest}
}
\description{
Compute the expected component for all models
}
\keyword{internal}
