% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinorOptimizationFunctions.R
\name{contain}
\alias{contain}
\title{Check whether one element is a subset of another element}
\usage{
contain(s1, s2)
}
\arguments{
\item{s1}{smaller subset}

\item{s2}{complete set}
}
\description{
Check whether one element is a subset of another element
}
\references{
This function is based on the "contain" function by Le and Singleton (2018). \cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
 Available at: https://cepr.org/40029
}
\keyword{internal}
