% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_K1Q.R
\name{Convert2JordanForm}
\alias{Convert2JordanForm}
\title{Convert a generic matrix to its Jordan form}
\usage{
Convert2JordanForm(K1XQ)
}
\arguments{
\item{K1XQ}{squared matrix in non-Jordan form}
}
\value{
squared matrix in Jordam form
}
\description{
Convert a generic matrix to its Jordan form
}
\details{
this Jordan matrix form handles real, complex and repeated eigenvalues.
}
\references{
This function is based on the "Convert2JordanForm" function by Le and Singleton (2018).\cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
 Available at: https://cepr.org/40029
}
\keyword{internal}
