% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliarBootstrapFunctions.R
\name{DataSet_BS}
\alias{DataSet_BS}
\title{Prepare the factor set for GVAR models  (Bootstrap version)}
\usage{
DataSet_BS(ModelType, RiskFactors, Wgvar, Economies, FactorLabels)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{RiskFactors}{Complete set of risk factors (KxT)}

\item{Wgvar}{transition matrix from GVAR models (CxC)}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}
}
\description{
Prepare the factor set for GVAR models  (Bootstrap version)
}
\keyword{internal}
