% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df__dx.R
\name{df__dx}
\alias{df__dx}
\title{Computes numerical first order derivative of f(x)}
\usage{
df__dx(f, x)
}
\arguments{
\item{f}{function which contains vector (J x T) valued function handle}

\item{x}{parameter values}
}
\value{
transformed matrix (MN x JT)
}
\description{
Computes numerical first order derivative of f(x)
}
\references{
This function is based on the "df__dx" function by Le and Singleton (2018). \cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
 Available at: https://cepr.org/40029
}
