% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{Est_K1h}
\alias{Est_K1h}
\title{Estimate K1h}
\usage{
Est_K1h(horiz, min_mat, max_mat, time_step, B_coef)
}
\arguments{
\item{horiz}{Minimum horizon observed}

\item{min_mat}{shortest maturity observed}

\item{max_mat}{longest maturity observed}

\item{time_step}{time unit of the model (scalar)}

\item{B_coef}{matrix of beta coefficients}
}
\description{
Estimate K1h
}
\keyword{internal}
