% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{Est_RestOLS}
\alias{Est_RestOLS}
\title{Estimate a restricted OLS model}
\usage{
Est_RestOLS(LHS, RHS, Rmat)
}
\arguments{
\item{LHS}{left hand side variables (M x T).}

\item{RHS}{right hand side variables (should include the intercept, if desired) (N x T).}

\item{Rmat}{matrix of constraints (M x N). Entries containing NAs are treated as free parameters.}
}
\value{
matrix of coefficient (M x N)
}
\description{
Estimate a restricted OLS model
}
\keyword{internal}
