% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{EstimationSigma_GVARrest}
\alias{EstimationSigma_GVARrest}
\title{Estimate numerically the variance-covariance matrix from the GVAR-based models}
\usage{
EstimationSigma_GVARrest(SigmaUnres, res, IdxVarRest)
}
\arguments{
\item{SigmaUnres}{Unrestricted variance-covariance matrix (K x K)}

\item{res}{residuals from the VAR of a GVAR model (K x T)}

\item{IdxVarRest}{index of the variable that is selected as strictly exogenous}
}
\value{
restricted version of the variance-covariance matrix a GVAR model (K x K)
}
\description{
Estimate numerically the variance-covariance matrix from the GVAR-based models
}
\keyword{internal}
