% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{ExpectedComponent}
\alias{ExpectedComponent}
\title{Get the expected component of all models}
\usage{
ExpectedComponent(
  ModelPara,
  InputsForOutputs,
  ModelType,
  Economies,
  FactorLabels,
  WishFP = FALSE
)
}
\arguments{
\item{ModelPara}{list of model parameter estimates}

\item{InputsForOutputs}{list containing the desired horizon of analysis for the model fit, IRFs, GIRFs, FEVDs, GFEVDs,
and risk premia decomposition}

\item{ModelType}{desired model type}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{FactorLabels}{string-list based which contains all the labels of all the variables present in the model}

\item{WishFP}{If users wants to compute the forward premia. Default is FALSE.}
}
\description{
Get the expected component of all models
}
\keyword{internal}
