% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBoundsBootstrap.R
\name{FactorBounds_IRFandGIRF}
\alias{FactorBounds_IRFandGIRF}
\title{Compute the confidence bounds for the model P-dynamics}
\usage{
FactorBounds_IRFandGIRF(
  ModelBootstrap,
  quants,
  ModelType,
  ndraws,
  Horiz,
  FacDim,
  LabIRF,
  Economies
)
}
\arguments{
\item{ModelBootstrap}{numerical output set from the bootstrap analysis}

\item{quants}{quantile of the confidence bounds}

\item{ModelType}{desired model type}

\item{ndraws}{number of draws}

\item{Horiz}{horizon of numerical outputs}

\item{FacDim}{dimension of the risk factor set}

\item{LabIRF}{vector containing the labels "IRF" and "GIRF"}

\item{Economies}{Economies that are part of the economic system}
}
\description{
Compute the confidence bounds for the model P-dynamics
}
\keyword{internal}
