% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{Factors_NonOrtho}
\alias{Factors_NonOrtho}
\title{Makes the pre-allocation of the factors set for JLL-based models}
\usage{
Factors_NonOrtho(NonOrthoFactors, JLLinputs, FactorLab, N)
}
\arguments{
\item{NonOrthoFactors}{Risk factors before the orthogonalization (FxT)}

\item{JLLinputs}{List of necessary inputs to estimate JLL-based setups}

\item{FactorLab}{Variable labels from JLL-based models}

\item{N}{number of country-specific spanned factors (scalar)}
}
\description{
Makes the pre-allocation of the factors set for JLL-based models
}
\keyword{internal}
