% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{FeedbackMatrixRestrictionsJLL}
\alias{FeedbackMatrixRestrictionsJLL}
\title{Set the zero-restrictions on the feedback matrix of JLL's P-dynamics}
\usage{
FeedbackMatrixRestrictionsJLL(DomUnit, K, G, M, N)
}
\arguments{
\item{DomUnit}{Name of the economy which is assigned as the dominant unit. \cr
If no dominant unit is assigned, then this variable is defined as "none"}

\item{K}{Total number of risk factors of the economic system (scalar)}

\item{G}{Number of global unspanned factors (scalar)}

\item{M}{Number of country-specific unspanned factors (scalar)}

\item{N}{Number of country-specific spanned factors (scalar)}
}
\value{
matrix containing the restrictions of the feedback matrix (K x K)
}
\description{
Set the zero-restrictions on the feedback matrix of JLL's P-dynamics
}
\keyword{internal}
