% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{FEVDandGFEVD}
\alias{FEVDandGFEVD}
\title{FEVDs and GFEVDs for all models}
\usage{
FEVDandGFEVD(ModelType, ModelPara, FEVDhoriz, FactorLabels, Economies)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelPara}{list of model parameter estimates (see the "Optimization" function)}

\item{FEVDhoriz}{single numerical vector containing the desired horizon of analysis for the FEVDs and GFEVDs}

\item{FactorLabels}{string-list based which contains all the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
FEVDs and GFEVDs for all models
}
\details{
Structural shocks are identified via Cholesky decomposition
}
\keyword{internal}
