% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBoundsBootstrap.R
\name{FEVDandGFEVDbs_sepQ}
\alias{FEVDandGFEVDbs_sepQ}
\title{Creates the confidence bounds and the graphs of FEVDs and GFEVDs after bootstrap ("sep Q" models)}
\usage{
FEVDandGFEVDbs_sepQ(
  ModelType,
  ModelBootstrap,
  NumOutPE,
  InputsForOutputs,
  Economies,
  PathsGraphs
)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelBootstrap}{list containing the complete set of model parameters after bootstrap estimation procedure}

\item{NumOutPE}{list of model parameter point estimates}

\item{InputsForOutputs}{list conataining the desired inputs for the construction of the outputs of interest}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{PathsGraphs}{path of the folder in which the graphs will be saved}
}
\description{
Creates the confidence bounds and the graphs of FEVDs and GFEVDs after bootstrap ("sep Q" models)
}
\keyword{internal}
