% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{FEVDandGFEVDs_Graphs}
\alias{FEVDandGFEVDs_Graphs}
\title{Generates graphs for FEVDs and GFEVDs}
\usage{
FEVDandGFEVDs_Graphs(OutputType, FEVDlist, nmVarInt, Lab_Fac, PathsGraphs)
}
\arguments{
\item{OutputType}{available options are "FEVD", "GFEVD", "FEVD Ortho" and "GFEVD Ortho"}

\item{FEVDlist}{list of FEVD and GFEVD outputs}

\item{nmVarInt}{name of variable of interest. Options: "Factors" and "Yields"}

\item{Lab_Fac}{label of the model factors}

\item{PathsGraphs}{Path to save graphs}
}
\description{
Generates graphs for FEVDs and GFEVDs
}
\keyword{internal}
