% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{FF}
\alias{FF}
\title{mean of the llk function used in the estimation of the selected ATSM}
\usage{
FF(x0, FFvectorized)
}
\arguments{
\item{x0}{vector of parameters to be estimated numerically}

\item{FFvectorized}{log-likelihood function}
}
\description{
mean of the llk function used in the estimation of the selected ATSM
}
\keyword{internal}
