% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{FFtemporary}
\alias{FFtemporary}
\title{Mean of the llk function used in the estimation of the selected ATSM}
\usage{
FFtemporary(xtemp, scaling_vector, FFvectorized)
}
\arguments{
\item{xtemp}{temporary vector of parameters to be estimated numerically}

\item{scaling_vector}{scaling factor}

\item{FFvectorized}{log-likelihood function}
}
\description{
Mean of the llk function used in the estimation of the selected ATSM
}
\keyword{internal}
