% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{Fit_Subplot}
\alias{Fit_Subplot}
\title{Build subplot for fitted yields}
\usage{
Fit_Subplot(
  YieldData,
  ModelFit,
  ModelImplied,
  MatLength,
  mat,
  Economies,
  ModelType,
  PathsGraphs
)
}
\arguments{
\item{YieldData}{Time series of bond yields}

\item{ModelFit}{Time series of fitted bond yields}

\item{ModelImplied}{Time series of model-implied bond yields}

\item{MatLength}{number of country-specific maturities}

\item{mat}{vector of maturities}

\item{Economies}{Economies of the economic system}

\item{ModelType}{Desired estimated model}

\item{PathsGraphs}{Path to save the graphs}
}
\description{
Build subplot for fitted yields
}
\keyword{internal}
