% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{FitgraphsSep}
\alias{FitgraphsSep}
\title{Model fit graphs for ("sep Q" models)}
\usage{
FitgraphsSep(
  ModelType,
  WishFitgraphs,
  ModelPara,
  NumOut,
  Economies,
  PathsGraphs
)
}
\arguments{
\item{ModelType}{a string-vector containing the label of the model to be estimated}

\item{WishFitgraphs}{binary variable: set 1, if the user wishes graphs to be generated; or set 0, otherwise}

\item{ModelPara}{List of model parameter estimates (See the "Optimization" function)}

\item{NumOut}{list of computed outputs containing the model fit, IRFs, FEVDs, GIRFs, and GFEVDs}

\item{Economies}{a string-vector containing the names of the economies which are part of the economic system}

\item{PathsGraphs}{Path of the folder in which the graphs will be saved}
}
\description{
Model fit graphs for ("sep Q" models)
}
