% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBoundsBootstrap.R
\name{FolderCreation_Boot}
\alias{FolderCreation_Boot}
\title{Creates folder to store graphs generated from the bootstrap analysis}
\usage{
FolderCreation_Boot(
  ModelType,
  LabIRF,
  Economies,
  OutType,
  Folder2save,
  Ortho = FALSE
)
}
\arguments{
\item{ModelType}{Desired model type}

\item{LabIRF}{Output types "IRF", "GIRF" and "IRF Ortho"}

\item{Economies}{economies of the economic system}

\item{OutType}{Available option "Factors" or "Yields}

\item{Folder2save}{Folder path where the outputs will be stored.}

\item{Ortho}{Option for orthogonal outputs, for JLL models. Default is FALSE.}
}
\description{
Creates folder to store graphs generated from the bootstrap analysis
}
\keyword{internal}
