% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Folder_Creation.R
\name{FolderCreationPoint}
\alias{FolderCreationPoint}
\title{Creates the folders and the path in which the graphical outputs are stored (ponit estimate version)}
\usage{
FolderCreationPoint(ModelType, Economies)
}
\arguments{
\item{ModelType}{a string-vector containing the label of the model to be estimated}

\item{Economies}{a string-vector containing the names of the economies which are part of the economic system}
}
\description{
Creates the folders and the path in which the graphical outputs are stored (ponit estimate version)
}
\keyword{internal}
