% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{FolderPrep_IRFs}
\alias{FolderPrep_IRFs}
\title{Create folders for storing IRFs and GIRFs}
\usage{
FolderPrep_IRFs(
  OutputType,
  WishPdynamicsgraphs,
  WishYieldsgraphs,
  Economies,
  ModelType
)
}
\arguments{
\item{OutputType}{available options are "IRF", "GIRF", "IRF Ortho" and "GIRF Ortho"}

\item{WishPdynamicsgraphs}{binary variable specifing whether the user whishes IRFs and/or GIRFs of risk factors}

\item{WishYieldsgraphs}{binary variable specifing whether the user whishes IRFs and/or GIRFs of bond yields}

\item{Economies}{Set of economies that are part of the economic system}

\item{ModelType}{Desired modem type}
}
\description{
Create folders for storing IRFs and GIRFs
}
\keyword{internal}
