% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{ForwardPremia}
\alias{ForwardPremia}
\title{Compute the forward premia for all models}
\usage{
ForwardPremia(
  ModelPara,
  avexpFP,
  ModelType,
  FactorLabels,
  InputsForOutputs,
  Economies
)
}
\arguments{
\item{ModelPara}{list of model parameter estimates}

\item{avexpFP}{list containing the country-specific expected component of the forward period}

\item{ModelType}{desired model type}

\item{FactorLabels}{List of factor labels}

\item{InputsForOutputs}{list containing the desired horizon of analysis for the model fit, IRFs, GIRFs, FEVDs, GFEVDs,
and risk premia decomposition}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Compute the forward premia for all models
}
\keyword{internal}
