% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionf.R
\name{Functionf}
\alias{Functionf}
\title{Set up the vector-valued objective function (Point estimate)}
\usage{
Functionf(MLEinputs, Economies, mat, DataFrequency, FactorLabels, ModelType)
}
\arguments{
\item{MLEinputs}{Set of inputs that are necessary to the log-likelihood function}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{mat}{vector of maturities (in years) of yields used in estimation (J x 1)}

\item{DataFrequency}{character-based vector: "Daily All Days", "Daily Business Days", "Weekly", "Monthly", "Quarterly", "Annually"}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{ModelType}{string-vector containing the label of the model to be estimated}
}
\value{
objective function
}
\description{
Set up the vector-valued objective function (Point estimate)
}
\examples{
# See examples in the vignette file of this package (Section 4).

}
