% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionf.R
\name{Functionf_Boot}
\alias{Functionf_Boot}
\title{Set up the vector-valued objective function (Bootstrap)}
\usage{
Functionf_Boot(
  ModelType,
  MLEinputsBS,
  Economies,
  mat,
  dt,
  FactorLabels,
  residBS,
  MaxEigen,
  JLLinputs,
  GVARinputs
)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{MLEinputsBS}{Set of inputs that are necessary to the log-likelihood function}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{mat}{vector of maturities (in years) of yields used in estimation (J x 1)}

\item{dt}{adjusted yearly frequency of the data}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{residBS}{indexes of the re-ordered bootstrap residuals}

\item{MaxEigen}{largest eigenvalue under the P-dynamics}

\item{JLLinputs}{necessary inputs for the estimation of JLL-based models}

\item{GVARinputs}{necessary inputs for the estimation of GVAR-based models}
}
\description{
Set up the vector-valued objective function (Bootstrap)
}
\keyword{internal}
