% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionML_vec.R
\name{FunctionML_vec}
\alias{FunctionML_vec}
\title{Use function ML to generate the outputs from a ATSM}
\usage{
FunctionML_vec(
  x,
  ML_fun,
  ListInputSet,
  ModelType,
  FactorLabels,
  Economies,
  JLLinputs,
  GVARinputs,
  WithEstimation
)
}
\arguments{
\item{x}{vector containing all the vectorized auxiliary parameters}

\item{ML_fun}{vector-valued objective function (function)}

\item{ListInputSet}{variable inputs used in the optimization (see inputs from "optimization" function)}

\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{JLLinputs}{Set of necessary inputs used in the estimation of the JLL-based models (see "JLL" function)}

\item{GVARinputs}{Set of necessary inputs used in the estimation of the GVAR-based models (see "GVAR" function)}

\item{WithEstimation}{if TRUE, returns only the values of the likelihood function, else generates the entire set of outputs}
}
\description{
Use function ML to generate the outputs from a ATSM
}
\keyword{internal}
