% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{Gather_Forecasts}
\alias{Gather_Forecasts}
\title{Gather several forecast dates}
\usage{
Gather_Forecasts(Forecast_OneDate, Forecast_AllDates, Economies, ModelType)
}
\arguments{
\item{Forecast_OneDate}{Bond yield forecasts for a single date}

\item{Forecast_AllDates}{Bond yield forecasts for multiple dates}

\item{Economies}{String-vector containing the names of the economies which are part of the economic system}

\item{ModelType}{String-vector containing the label of the model to be estimated}
}
\description{
Gather several forecast dates
}
\keyword{internal}
