% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{GaussianDensity}
\alias{GaussianDensity}
\title{computes the density function of a gaussian process}
\usage{
GaussianDensity(res, SS)
}
\arguments{
\item{res}{matrix of residuals (N x T)}

\item{SS}{covariance matrices or array of covariance matrices (N x N)}
}
\value{
y vector of density (1 x T)
}
\description{
computes the density function of a gaussian process
}
\keyword{internal}
