% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{Gen_art_series}
\alias{Gen_art_series}
\title{Simulate N_Boot dataset from the P-dynamics}
\usage{
Gen_art_series(K1Z_j, N_Boot, RFs)
}
\arguments{
\item{K1Z_j}{Feedback matrix at the j_th iteration}

\item{N_Boot}{Number of bootstrap samples per iteration.}

\item{RFs}{A numeric matrix (T x F) representing the time series of risk factors.}
}
\description{
Simulate N_Boot dataset from the P-dynamics
}
\keyword{internal}
