% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{genVARbrw}
\alias{genVARbrw}
\title{Generate M data sets from VAR(1) model}
\usage{
genVARbrw(Phi, M, RiskFactors)
}
\arguments{
\item{Phi}{feedback matrix (F x F)}

\item{M}{number of Monte Carlo replications}

\item{RiskFactors}{time series of the risk factors (T x F)}
}
\description{
Generate M data sets from VAR(1) model
}
\references{
Bauer, Rudebusch and, Wu (2012). "Correcting Estimation Bias in Dynamic Term Structure Models". \cr
This function is similar to the"genVAR" Matlab function available at Cynthia Wu's website
(https://sites.google.com/view/jingcynthiawu/).
}
