% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A0N__BnAn.R
\name{Get__BnXAnX}
\alias{Get__BnXAnX}
\title{Compute the cross-section loadings of yields of a canonical A0_N model}
\usage{
Get__BnXAnX(mat, K1XQ, ModelType, r0 = NULL, SSX = NULL, Economies)
}
\arguments{
\item{mat}{vector of maturities (J x 1).}

\item{K1XQ}{risk-neutral feedback matrix (N x N)}

\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{r0}{the long run risk neutral (scalar)}

\item{SSX}{covariance matrix of the latent states (N x N)}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Compute the cross-section loadings of yields of a canonical A0_N model
}
\references{
\itemize{
 \item Dai, Q. and Singleton, K. (2000). "Specification Analysis of Affine Term Structure Models". The Journal of Finance.
 \item Joslin, S., Singleton, K. and Zhu, H. (2011). "A new perspective on Gaussian dynamic term structure models".
 The Review of Financial Studies.
 }
}
\keyword{internal}
