% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{Get_a0}
\alias{Get_a0}
\title{Obtain the country-specific a0}
\usage{
Get_a0(GVARinputs, ParaVARX)
}
\arguments{
\item{GVARinputs}{List of inputs for GVAR-based models}

\item{ParaVARX}{List containing the set of VARX model parameters}
}
\description{
Obtain the country-specific a0
}
\keyword{internal}
