% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{Get_As}
\alias{Get_As}
\title{Compute the A loadings}
\usage{
Get_As(LoadBs, Wpca, r0, dt, Economies, ModelType)
}
\arguments{
\item{LoadBs}{list containing the B loadings}

\item{Wpca}{matrix of weights of the portfolios observed without errors (N x J or CN x J)}

\item{r0}{long-run interest rate (scalar or vector with length C)}

\item{dt}{time interval unit of the model (scalar). For instance, if data is (i) monthly, dt <- 12; (ii) quarterly, dt <- 4; (iii) yearly, dt <- 1}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{ModelType}{string-vector containing the label of the model to be estimated}
}
\description{
Compute the A loadings
}
\keyword{internal}
