% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{Get_Bs}
\alias{Get_Bs}
\title{BUild the B loadings}
\usage{
Get_Bs(mat, dt, K1XQ, SSZ, Wpca, FactorLabels, Economy, ModelType)
}
\arguments{
\item{mat}{vector of maturities (in years) of yields used in estimation (J x 1)}

\item{dt}{time interval unit of the model (scalar). For instance, if data is (i) monthly, dt <- 12; (ii) quarterly, dt <- 4; (iii) yearly, dt <- 1}

\item{K1XQ}{risk-neutral feedback matrix (N x N or CN x CN)}

\item{SSZ}{variance-covariance matrix (F x F)}

\item{Wpca}{matrix of weights of the portfolios observed without errors (N x J or CN x J)}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economy}{string-vector containing the names of the economies which are part of the economic system}

\item{ModelType}{string-vector containing the label of the model to be estimated}
}
\description{
BUild the B loadings
}
\keyword{internal}
