% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{Get_FeedMat_NoBC}
\alias{Get_FeedMat_NoBC}
\title{Estimate feedback matrix from several models (No bias-corrected version)}
\usage{
Get_FeedMat_NoBC(
  RiskFactors,
  ModelType,
  Economies,
  GVARinputs,
  JLLinputs,
  FactorLabels,
  CheckInpts = FALSE
)
}
\arguments{
\item{RiskFactors}{A numeric matrix (T x F) representing the time series of risk factors.}

\item{Economies}{A character vector containing the names of the economies included in the system.}

\item{GVARinputs}{List. Inputs for GVAR model estimation.}

\item{JLLinputs}{List. Inputs for JLL model estimation.}

\item{FactorLabels}{A list of character vectors with labels for all variables in the model.}

\item{CheckInpts}{Check for the consistency of the provided inputs. Default is FALSE.}
}
\description{
Estimate feedback matrix from several models (No bias-corrected version)
}
\keyword{internal}
