% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{Get_G0G1Sigma}
\alias{Get_G0G1Sigma}
\title{Get the intercept, feedback matrix and the variance-covariance matrix from GVAR without global factors}
\usage{
Get_G0G1Sigma(ParaVARX, GVARinputs, Ai0, Ai1, Wi)
}
\arguments{
\item{ParaVARX}{Set of VARX model parameters}

\item{GVARinputs}{List of inputs for GVAR-based models}

\item{Ai0}{list containing the country-specific intercepts}

\item{Ai1}{list containing the country-specific feedback matrices}

\item{Wi}{list containing the country-specific link matrices}
}
\description{
Get the intercept, feedback matrix and the variance-covariance matrix from GVAR without global factors
}
\keyword{internal}
