% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{Get_Gy1}
\alias{Get_Gy1}
\title{Compute the feedback matrix from a GVAR model with global factors}
\usage{
Get_Gy1(ParaVARX, GVARinputs, G1, Phi.w1)
}
\arguments{
\item{ParaVARX}{List containing the set of VARX model parameters}

\item{GVARinputs}{List of inputs for GVAR-based models}

\item{G1}{feedback matrix from a GVAR without global variables}

\item{Phi.w1}{feedback matrix from a marginal model}
}
\description{
Compute the feedback matrix from a GVAR model with global factors
}
\keyword{internal}
