% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{Get_Sigama_JLL}
\alias{Get_Sigama_JLL}
\title{Compute Sigmas/Cholesky factorizations}
\usage{
Get_Sigama_JLL(JLLinputs, FacSet, Para_Ortho_Reg, Para_Ortho_VAR, N)
}
\arguments{
\item{JLLinputs}{List of necessary inputs to estimate JLL-based setups}

\item{FacSet}{Set of factors used in the estimation of JLL-based setups}

\item{Para_Ortho_Reg}{Set of parameters obtained from the JLL regressions}

\item{Para_Ortho_VAR}{Set of parameters obtained from the VAR(1) of JLL-based models}

\item{N}{number of country-specific spanned factors (scalar)}
}
\description{
Compute Sigmas/Cholesky factorizations
}
\keyword{internal}
