% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{Get_SigmaYields}
\alias{Get_SigmaYields}
\title{Compute the variance-covariance matrix of the bond yields}
\usage{
Get_SigmaYields(YieldsTS, N, mat, WpcaFull, se, ModelType)
}
\arguments{
\item{YieldsTS}{matrix of yields used in estimation  (J x T or CJ x T)}

\item{N}{number of country-specific spanned factors}

\item{mat}{vector of maturities (in years) of yields used in estimation (J x 1)}

\item{WpcaFull}{composite matrix of weights the portfolios observed with and without errors}

\item{se}{Variance of the portfolio of yields observed with error (scalar). Default is set to NULL}

\item{ModelType}{string-vector containing the label of the model to be estimated}
}
\description{
Compute the variance-covariance matrix of the bond yields
}
\keyword{internal}
