% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{GetAuxPara}
\alias{GetAuxPara}
\title{Compute the auxiliary parameters a.}
\usage{
GetAuxPara(
  ParaValue,
  Const_Type_Full,
  Economies,
  FactorLabels,
  JLLinputs = NULL
)
}
\arguments{
\item{ParaValue}{Parameter value}

\item{Const_Type_Full}{character-based vector that describes the constraints. Constraints are:
\itemize{
\item 'Jordan' for single-country models;
\item 'Jordan; stationary' for single-country models;
\item 'Jordan MultiCountry' for multicountry models;
\item 'Jordan MultiCountry; stationary' for multicountry models;
\item 'psd' for JPS-based models;
\item 'BlockDiag' for GVAR-based models;
\item 'JLLstructure' for JLL-based models;
  }}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{JLLinputs}{list of necessary inputs for the estimation of JLL-based models (see "JLL" function)}
}
\description{
Compute the auxiliary parameters a.
}
\keyword{internal}
