% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{Getdt}
\alias{Getdt}
\title{Get delta t}
\usage{
Getdt(DataFrequency)
}
\arguments{
\item{DataFrequency}{single element character-based vector. Available options are: "Daily All Days", \cr
"Daily Business Days", "Weekly", "Monthly",  "Quarterly", "Annually"}
}
\description{
Get delta t
}
\keyword{internal}
