% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{GetLabels_JLL}
\alias{GetLabels_JLL}
\title{Generate the variable labels of the JLL models}
\usage{
GetLabels_JLL(NonOrthoFactors, JLLinputs, G)
}
\arguments{
\item{NonOrthoFactors}{Risk factors before the orthogonalization (FxT)}

\item{JLLinputs}{List of necessary inputs to estimate JLL-based setups}

\item{G}{number of global unspanned factors (scalar)}
}
\description{
Generate the variable labels of the JLL models
}
\keyword{internal}
