% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{GetLabels_sepQ}
\alias{GetLabels_sepQ}
\title{Generate the factor labels for models estimated on a country-by-country bases}
\usage{
GetLabels_sepQ(Economy, ModelType, FactorLabels)
}
\arguments{
\item{Economy}{string containing the names of the economy to be estimated}

\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{FactorLabels}{list containing the factor labels}
}
\description{
Generate the factor labels for models estimated on a country-by-country bases
}
\keyword{internal}
