% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionf_vectorized.R
\name{getpara}
\alias{getpara}
\title{Extract the parameter values from the parameter list set}
\usage{
getpara(ListInputSet)
}
\arguments{
\item{ListInputSet}{All parameter features}
}
\description{
Extract the parameter values from the parameter list set
}
\references{
This function is modified version of the "getpara" function by Le and Singleton (2018). \cr
"A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
(Euro Area Business Cycle Network Training School - Term Structure Modelling).
Available at: https://cepr.org/40029
}
\keyword{internal}
