% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{GetPdynPara_NoBC}
\alias{GetPdynPara_NoBC}
\title{Compute P-dynamics parameters without using the bias correction method from BRW (2012)}
\usage{
GetPdynPara_NoBC(
  ModelType,
  RiskFactors,
  Economies,
  N,
  GVARinputs,
  JLLinputs,
  CheckInpts = F
)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{RiskFactors}{time series of risk factors (F x T). Could be stored in a list depending on the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{N}{number of country-specific spanned factors}

\item{GVARinputs}{list of necessary inputs for the estimation of GVAR-based models (see "GVAR" function)}

\item{JLLinputs}{list of necessary inputs for the estimation of JLL-based models (see "JLL" function)}
}
\description{
Compute P-dynamics parameters without using the bias correction method from BRW (2012)
}
\keyword{internal}
