% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{GetYields_AllCountries}
\alias{GetYields_AllCountries}
\title{Gather all country-specific yields in a single matrix of dimension CJ x T}
\usage{
GetYields_AllCountries(ModelPara, Economies, ModelType)
}
\arguments{
\item{ModelPara}{List of model parameter estimates}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{ModelType}{a string-vector containing the label of the model to be estimated}
}
\description{
Gather all country-specific yields in a single matrix of dimension CJ x T
}
\keyword{internal}
