% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{GFEVDgraphsJoint}
\alias{GFEVDgraphsJoint}
\title{GFEVDs graphs for "joint Q" models}
\usage{
GFEVDgraphsJoint(
  ModelType,
  NumOut,
  WishPdynamicsgraphs,
  WishYieldsgraphs,
  GFEVDhoriz,
  PathsGraphs
)
}
\arguments{
\item{ModelType}{a string-vector containing the label of the model to be estimated}

\item{NumOut}{list of computed outputs containing the model fit, IRFs, FEVDs, GIRFs, and GFEVDs}

\item{WishPdynamicsgraphs}{binary variable: set 1, if the user wishes graphs to be generated; or set 0, otherwise}

\item{WishYieldsgraphs}{binary variable: set 1, if the user wishes graphs to be generated; or set 0, otherwise}

\item{GFEVDhoriz}{single numerical vector conataining the desired horizon of analysis for the GFEVDs}

\item{PathsGraphs}{Path of the folder in which the graphs will be saved}
}
\description{
GFEVDs graphs for "joint Q" models
}
\keyword{internal}
