% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputsBootstrap.R
\name{GIRFjointOrthoJLL_BS}
\alias{GIRFjointOrthoJLL_BS}
\title{GIRFs after bootstrap for JLL-based models}
\usage{
GIRFjointOrthoJLL_BS(
  ModelType,
  ModelParaBoot,
  GIRFhoriz,
  FactorLabels,
  Economies
)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelParaBoot}{list of model parameter estimates (see the "Optimization" function) after a bootstrap draw}

\item{GIRFhoriz}{single numerical vector conataining the desired horizon of analysis for the GIRFs}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
GIRFs after bootstrap for JLL-based models
}
\references{
\itemize{
\item This function is a modified and extended version of the "irf" function from
Smith, L.V. and A. Galesi (2014). GVAR Toolbox 2.0, available at https://sites.google.com/site/gvarmodelling/gvar-toolbox.

\item Pesaran and Shin, 1998. "Generalized impulse response analysis in linear multivariate models" (Economics Letters)
}
}
