% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_GlobalMacro_2023.R
\docType{data}
\name{GlobalMacro}
\alias{GlobalMacro}
\title{Data: Risk Factors - Candelon and Moura (2023)}
\format{
matrix containing the  risk factors of the models
}
\usage{
data("CM_GlobalMacro_2023")
}
\description{
Risk factors data used in Candelon and Moura (2023)
}
\references{
Candelon, B. and Moura, R. (2023) "Sovereign yield curves and the COVID-19 in emerging markets". (Economic Modelling)
}
\keyword{Factors}
\keyword{Risk}
