% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_GlobalMacroFactors.R
\docType{data}
\name{GlobalMacroVar}
\alias{GlobalMacroVar}
\title{Data: Risk Factors - Candelon and Moura (2024, JFEC)}
\format{
matrix containing the  risk factors of the models
}
\usage{
data("CM_GlobalMacroFactors")
}
\description{
Risk factors data used in Candelon and Moura (2024, JFEC)
}
\references{
Candelon, B. and Moura, R. (2024) "A Multicountry Model of the Term Structures of Interest Rates with a GVAR". (Journal of Financial Econometrics)
}
\keyword{Factors}
\keyword{Risk}
