% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{GraphicalOutputs}
\alias{GraphicalOutputs}
\title{Generate the graphical outputs for the selected models (Point estimate)}
\usage{
GraphicalOutputs(
  ModelType,
  ModelPara,
  NumOut,
  InputsForOutputs,
  Economies,
  FactorLabels
)
}
\arguments{
\item{ModelType}{a string-vector containing the label of the model to be estimated}

\item{ModelPara}{List of model parameter estimates (See the "Optimization" function)}

\item{NumOut}{list of computed outputs containing the model fit, IRFs, FEVDs, GIRFs, and GFEVDs}

\item{InputsForOutputs}{list containing the desired inputs for the construction of the desired output}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}
}
\description{
Generate the graphical outputs for the selected models (Point estimate)
}
