% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{GVAR_PrepFactors}
\alias{GVAR_PrepFactors}
\title{Prepare risk factors for the estimation of the GVAR model}
\usage{
GVAR_PrepFactors(GVARinputs, DomLabels, StarLabels, GlobalLabels, N)
}
\arguments{
\item{GVARinputs}{List of inputs for GVAR-based models}

\item{DomLabels}{string-based vector containing label of the domestic risk factors}

\item{StarLabels}{string-based vector containing label of the star domestic risk factors}

\item{GlobalLabels}{string-based vector containing label of the global risk factors}

\item{N}{number of country-specific spanned factors (scalar)}
}
\description{
Prepare risk factors for the estimation of the GVAR model
}
\keyword{internal}
