% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{Idx_UnspanFact}
\alias{Idx_UnspanFact}
\title{Obtain the indexes of both the domestic and global unspanned factors}
\usage{
Idx_UnspanFact(RiskFactors_TS, FactorLabels, Economies)
}
\arguments{
\item{RiskFactors_TS}{Time series of risk factors for the jointly estimated models (CJ x T)}

\item{FactorLabels}{A string-list based which contains all the labels of all the variables present in the model}

\item{Economies}{String-vector containing the names of the economies which are part of the economic system}
}
\description{
Obtain the indexes of both the domestic and global unspanned factors
}
\keyword{internal}
