% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{IdxAllSpanned}
\alias{IdxAllSpanned}
\title{Find the indexes of the spanned factors}
\usage{
IdxAllSpanned(ModelType, FactorLabels, Economies)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Find the indexes of the spanned factors
}
