% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{IdxSpanned}
\alias{IdxSpanned}
\title{Extract the indexes related to the spanned factors in the variance-covariance matrix}
\usage{
IdxSpanned(G, M, N, C)
}
\arguments{
\item{G}{number of global unspanned factors (scalar)}

\item{M}{number of domestic unspanned factors per country (scalar)}

\item{N}{number of domestic spanned factors per country (scalar)}

\item{C}{number of countries of the economic system (scalar)}
}
\description{
Extract the indexes related to the spanned factors in the variance-covariance matrix
}
