% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{IDXZeroRestrictionsJLLVarCovOrtho}
\alias{IDXZeroRestrictionsJLLVarCovOrtho}
\title{Find the indexes of zero-restrictions from the orthogonalized variance-covariance matrix from the JLL-based models}
\usage{
IDXZeroRestrictionsJLLVarCovOrtho(M, N, G, Economies, DomUnit)
}
\arguments{
\item{M}{number of country-specific unspanned factors (scalar)}

\item{N}{number of country-specific spanned factors (scalar)}

\item{G}{number of global unspanned factors (scalar)}

\item{Economies}{Set of economies that are part of the economic system (string-vector)}

\item{DomUnit}{Name of the economy which is assigned as the dominant unit. \cr
If no dominant unit is assigned, then this variable is defined as "None"}
}
\value{
restricted version of the JLL of the Cholesky factorization (F x F)
}
\description{
Find the indexes of zero-restrictions from the orthogonalized variance-covariance matrix from the JLL-based models
}
\keyword{internal}
