% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{ImposeStat_Aux}
\alias{ImposeStat_Aux}
\title{Impose stationary constraint under the risk-neutral measure}
\usage{
ImposeStat_Aux(yy)
}
\arguments{
\item{yy}{numerical vector before imposing stationary constraint}
}
\description{
Impose stationary constraint under the risk-neutral measure
}
\keyword{internal}
