% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_InpOut.R
\docType{data}
\name{InpForOutEx}
\alias{InpForOutEx}
\title{Example of list inputs used in the construction of several model outputs}
\format{
list of inputs
}
\usage{
data("InpForOutEx")
}
\description{
List of inputs of a JPS-based model for Brazilian data
}
\keyword{-}
\keyword{Brazil}
\keyword{JPS}
\keyword{model}
