% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{Intra_Yields}
\alias{Intra_Yields}
\title{Fit the cross-section of yields using spline}
\usage{
Intra_Yields(Yields, mat_vec, time_step)
}
\arguments{
\item{Yields}{matrix containing country-specific yields  (J x T)}

\item{mat_vec}{vector of maturities expressed in years (J x 1)}

\item{time_step}{time unit of the model (scalar)}
}
\description{
Fit the cross-section of yields using spline
}
\keyword{internal}
