% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{InvMat_Robust}
\alias{InvMat_Robust}
\title{Robust method for matrix inversion}
\usage{
InvMat_Robust(M)
}
\arguments{
\item{M}{squared matrix}
}
\description{
Robust method for matrix inversion
}
\keyword{internal}
