% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{IRFandGIRF}
\alias{IRFandGIRF}
\title{IRFs and GIRFs for all models}
\usage{
IRFandGIRF(ModelType, ModelPara, IRFhoriz, FactorLabels, Economies)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelPara}{list of model parameter estimates (See the "Optimization" function)}

\item{IRFhoriz}{single numerical vector containing the desired horizon of analysis for the IRFs}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
IRFs and GIRFs for all models
}
\details{
The Structural shocks from the IRFs are identified via Cholesky decomposition
}
\keyword{internal}
