% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputsBootstrap.R
\name{IRFandGIRF_BS}
\alias{IRFandGIRF_BS}
\title{IRFs and GIRFs after bootstrap for all models}
\usage{
IRFandGIRF_BS(ModelType, ModelParaBoot, IRFhoriz, FactorLabels, Economies)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelParaBoot}{list of model parameter estimates (see the "Optimization" function) after a bootstrap draw}

\item{IRFhoriz}{single numerical vector containing the desired horizon of analysis for the IRFs}

\item{FactorLabels}{string-list based which contains all the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
IRFs and GIRFs after bootstrap for all models
}
\keyword{internal}
