% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{IRFandGIRFs_Format_Fac}
\alias{IRFandGIRFs_Format_Fac}
\title{Gather data for IRFs and GIRFs grahs (version "Factors")}
\usage{
IRFandGIRFs_Format_Fac(IRFFac)
}
\arguments{
\item{IRFFac}{Data-frame with basic features of a single IRF for risk factors}
}
\description{
Gather data for IRFs and GIRFs grahs (version "Factors")
}
\keyword{internal}
