% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{IRFandGIRFs_Format_Yields}
\alias{IRFandGIRFs_Format_Yields}
\title{Gather data for IRFs and GIRFs grahs (version "Yields")}
\usage{
IRFandGIRFs_Format_Yields(IRFYields)
}
\arguments{
\item{IRFYields}{Data-frame with basic features of a single IRF for yields}
}
\description{
Gather data for IRFs and GIRFs grahs (version "Yields")
}
\keyword{internal}
