% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{IRFjointOrthoJLL}
\alias{IRFjointOrthoJLL}
\title{Orthogonalized IRFs for JLL models}
\usage{
IRFjointOrthoJLL(ModelType, ModelPara, IRFhoriz, FactorLabels, Economies)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelPara}{list of model parameter estimates (see the "Optimization" function)}

\item{IRFhoriz}{single numerical vector conataining the desired horizon of analysis for the IRFs}

\item{FactorLabels}{string-list based which contains all the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Orthogonalized IRFs for JLL models
}
\details{
Structural shocks are identified via Cholesky decomposition
}
\keyword{internal}
