% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{Jac_approx}
\alias{Jac_approx}
\title{Main Jacobian approximation}
\usage{
Jac_approx(f, x, nlevels = 5L)
}
\arguments{
\item{f}{function to differentiate}

\item{x}{evaluation point}

\item{nlevels}{depth of extrapolation. Default is 5.}
}
\description{
Main Jacobian approximation
}
\references{
Le, A., & Singleton, K. J. (2018). Small Package of Matlab Routines for
Estimation of Some Term Structure Models. EABCN Training School.\cr
This function offers an independent R implementation that is informed
by the conceptual framework outlined in Le and Singleton (2018), but adapted to the
present modeling context.
}
\keyword{internal}
