% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buildx_xvec.R
\name{Jordan_JLL}
\alias{Jordan_JLL}
\title{Check for JLL models for Jordan restrictions (auxiliary form)}
\usage{
Jordan_JLL(ParaValue, C, N)
}
\arguments{
\item{ParaValue}{Constrained parameter value}

\item{C}{number of countries of the economic system}

\item{N}{number of country-specific spanned factors}
}
\description{
Check for JLL models for Jordan restrictions (auxiliary form)
}
\keyword{internal}
