% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getx.R
\name{killa}
\alias{killa}
\title{Eliminates the @}
\usage{
killa(s)
}
\arguments{
\item{s}{text vector containing the feature of the variable}
}
\description{
Eliminates the @
}
\references{
This function is a modified version of the "killa" function by Le and Singleton (2018). \cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
}
