% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabFac.R
\name{LabelsSpanned}
\alias{LabelsSpanned}
\title{Generate the labels of the spanned factors}
\usage{
LabelsSpanned(N)
}
\arguments{
\item{N}{number of spanned factors}
}
\description{
Generate the labels of the spanned factors
}
\keyword{internal}
