% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabFac.R
\name{LabelsStar}
\alias{LabelsStar}
\title{Generate the labels of the star variables}
\usage{
LabelsStar(FactorLabels)
}
\arguments{
\item{FactorLabels}{Factor labels}
}
\description{
Generate the labels of the star variables
}
