% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{llk_JLL_Sigma}
\alias{llk_JLL_Sigma}
\title{Build the log-likelihood function of the P-dynamics from the JLL-based models}
\usage{
llk_JLL_Sigma(VecPara, res, IdxNONzero, K)
}
\arguments{
\item{VecPara}{vector that contains all the non-zero entries of the lower-triangular part of the Cholesky factorization}

\item{res}{residuals from the VAR of the JLL model (K x T)}

\item{IdxNONzero}{vector that contains indexes of the matrix of the non-zero entries of the Cholesky factorization}

\item{K}{dimensions of the variance-covariance matrix (scalar)}
}
\value{
value of the log-likelihood function (scalar)
}
\description{
Build the log-likelihood function of the P-dynamics from the JLL-based models
}
\keyword{internal}
