% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataForEstimation.R
\name{Load_Excel_Data}
\alias{Load_Excel_Data}
\title{Read data from Excel files and transform them into a dataframe}
\usage{
Load_Excel_Data(ExcelFilePath)
}
\arguments{
\item{ExcelFilePath}{Path of the excel file}
}
\description{
Read data from Excel files and transform them into a dataframe
}
\examples{
 if (!requireNamespace("readxl", quietly = TRUE)) {
 stop(
   "Please install package \"readxl\" to use this feature.",
   call. = FALSE
 )

Load_Excel_Data(system.file("extdata", "MacroData.xlsx", package = "MultiATSM"))
Load_Excel_Data(system.file("extdata", "YieldsData.xlsx", package = "MultiATSM"))
 }
}
