% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{LoadData}
\alias{LoadData}
\title{Loads data sets from several papers}
\usage{
LoadData(DataPaper)
}
\arguments{
\item{DataPaper}{Available options are \code{BR_2017} (Bauer and Rudebusch, 2017) , \code{CM_2023} (Candelon and Moura, 2023), \code{CM_2024} (Candelon and Moura, 2024)}
}
\value{
Complete set of data from several papers.
}
\description{
Loads data sets from several papers
}
\examples{
# Example 1:
LoadData("BR_2017")

# Example 2:
LoadData("CM_2023")

# Example 3:
LoadData("CM_2024")

}
\references{
\enumerate{
\item Bauer and Rudebusch (2017). "Resolving the Spanning Puzzle in Macro-Finance Term Structure Models" (Review of Finance)
\item Candelon and Moura (2023). "Sovereign yield curves and the COVID-19 in emerging markets" (Economic Modelling)
\item Candelon and Moura (2024). "A Multicountry Model of the Term Structures of Interest Rates with a GVAR" (Journal of Financial Econometrics)
}
}
