% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{MarginalModelPara}
\alias{MarginalModelPara}
\title{Estimate the marginal model for the global factors}
\usage{
MarginalModelPara(GVARinputs)
}
\arguments{
\item{GVARinputs}{List of inputs for GVAR-based models}
}
\description{
Estimate the marginal model for the global factors
}
\keyword{internal}
