% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{MatAdjusted}
\alias{MatAdjusted}
\title{Adjust vector of maturities}
\usage{
MatAdjusted(mat, UnitYields)
}
\arguments{
\item{mat}{vector of maturities (J x 1)}

\item{UnitYields}{Available options: "Month" and "Year"}
}
\description{
Adjust vector of maturities
}
\keyword{internal}
