% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maturities.R
\name{Maturities}
\alias{Maturities}
\title{Create a vector of numerical maturities in years}
\usage{
Maturities(DataYields, Economies, UnitYields)
}
\arguments{
\item{DataYields}{matrix containing all yields of the system (JxT,if the model is single-country-based
or CJxT if the model is multy-country-based)}

\item{Economies}{vector containing names of all the economies of the system}

\item{UnitYields}{(i) "Month": if maturity of yields are expressed in months or
(ii) "Year": if maturity of yields are expressed in years}
}
\value{
Vector containing all observed maturities expressed in years
}
\description{
Create a vector of numerical maturities in years
}
\examples{
data('CM_Yields')
Economies <- c("China", "Brazil", "Mexico", "Uruguay")
Maturities(Yields, Economies, "Month")
}
