% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{MergeFEVD_graphs}
\alias{MergeFEVD_graphs}
\title{Limit the number of categories in FEVDs and GFEVDs graphs by merging small groups into other}
\usage{
MergeFEVD_graphs(FEVDlist, Threshold = 0.05)
}
\arguments{
\item{FEVDlist}{list of FEVD and GFEVD outputs}

\item{Threshold}{Threshold for merging factors. Default is 0.05.}
}
\description{
Limit the number of categories in FEVDs and GFEVDs graphs by merging small groups into other
}
\keyword{internal}
