% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{ML_stable}
\alias{ML_stable}
\title{Prevents algorithm to end up in ill-defined likelihood}
\usage{
ML_stable(par, MLvec)
}
\arguments{
\item{par}{vector of parameters}

\item{MLvec}{objective function}
}
\description{
Prevents algorithm to end up in ill-defined likelihood
}
\keyword{internal}
