% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{MLFunction}
\alias{MLFunction}
\title{Set up the vector-valued objective function (Point estimate)}
\usage{
MLFunction(
  MLEinputs,
  Economies,
  DataFrequency,
  FactorLabels,
  ModelType,
  BS_outType = FALSE
)
}
\arguments{
\item{MLEinputs}{Set of inputs that are necessary to the log-likelihood function}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{DataFrequency}{character-based vector: "Daily All Days", "Daily Business Days", "Weekly", "Monthly", "Quarterly", "Annually"}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{BS_outType}{Generates simplified output list in the bootstrap setting. Default is set to FALSE.}
}
\value{
objective function
}
\description{
Set up the vector-valued objective function (Point estimate)
}
\keyword{internal}
