% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{MLtemporary}
\alias{MLtemporary}
\title{Mean of the llk function used in the estimation of the selected ATSM}
\usage{
MLtemporary(xtemp, scaling_vector, MLvec)
}
\arguments{
\item{xtemp}{temporary vector of parameters to be estimated numerically}

\item{scaling_vector}{scaling factor}

\item{MLvec}{log-likelihood function}
}
\description{
Mean of the llk function used in the estimation of the selected ATSM
}
\keyword{internal}
