% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_repJPS.R
\docType{data}
\name{ModelPara}
\alias{ModelPara}
\title{Replications of the JPS (2014) outputs by the MultiATSM package}
\format{
list of inputs and outputs
\describe{
   \item{inputs}{general model inputs}
   \item{ests}{model parameters estimates (JPS form)}
   \item{llk}{log-likelihood of the observations}
   \item{rot}{model parameters estimates (rotation form)}
  }
}
\usage{
data("JPSrep")
}
\description{
Unspanned macro risk model outputs by the MultiATSM package
}
\keyword{(2014)}
\keyword{JPS}
\keyword{outputs}
