% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_Para.R
\docType{data}
\name{ModelParaEx}
\alias{ModelParaEx}
\title{Example of parameter set after optimization}
\format{
list of inputs and outputs
\describe{
   \item{inputs}{general model inputs}
   \item{ests}{model parameters estimates (JPS form)}
   \item{llk}{log-likelihood of the observations}
   \item{rot}{model parameters estimates (rotation form)}
  }
}
\usage{
data("ParaSetEx")
}
\description{
JPS-based model for Brazilian data
}
\keyword{-}
\keyword{Brazil}
\keyword{JPS}
\keyword{model}
