% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{mult__prod}
\alias{mult__prod}
\title{Efficient computation of matrix product for arrays}
\usage{
mult__prod(a, b)
}
\arguments{
\item{a}{array}

\item{b}{array}
}
\description{
Efficient computation of matrix product for arrays
}
\references{
This function is a simplified version of the "mult__prod" function by Le and Singleton (2018). \cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models."\cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
 Available at: https://cepr.org/40029
}
\keyword{internal}
