% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiATSM.R
\docType{package}
\name{MultiATSM}
\alias{MultiATSM}
\alias{MultiATSM-package}
\title{ATSM Package}
\description{
Estimation of several classes of affine term structure of interest rates models.
}
\author{
Rubens Moura \email{rubens.gtmoura@gmail.com}
}
