% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{multiprod_2terms}
\alias{multiprod_2terms}
\title{computes matrix product for arrays a and b:   c[,,i] = a[,,i] b[,,i]}
\usage{
multiprod_2terms(a, b)
}
\arguments{
\item{a}{array (M x N x T)}

\item{b}{array (N x K x T)}
}
\value{
c  array (M x K x T)
}
\description{
computes matrix product for arrays a and b:   c[,,i] = a[,,i] b[,,i]
}
\references{
This function is based on the "multiprod_2terms" function by Le and Singleton (2018).\cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
 Available at: https://cepr.org/40029
}
\keyword{internal}
