% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{NoOrthoVAR_JLL}
\alias{NoOrthoVAR_JLL}
\title{Obtain the non-orthogonalized model parameters}
\usage{
NoOrthoVAR_JLL(Para_Ortho_Reg, Para_Ortho_VAR)
}
\arguments{
\item{Para_Ortho_Reg}{Set of parameters obtained from the JLL regressions}

\item{Para_Ortho_VAR}{Set of parameters obtained from the VAR(1) of JLL-based models}
}
\description{
Obtain the non-orthogonalized model parameters
}
\keyword{internal}
