% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_NumOut.R
\docType{data}
\name{NumOutEx}
\alias{NumOutEx}
\title{Example of computed numerical outputs}
\format{
list of inputs and outputs
\describe{
   \item{PC var explained}{variance explained per spanned factor}
   \item{Fit}{model fit of bond yields}
   \item{IRF}{Impulse response function}
   \item{FEVD}{Forecast error variance decomposition}
   \item{GIRF}{Generalized impulse response function}
   \item{GFEVD}{Generalized Forecast error variance decomposition}
   \item{TermPremiaDecomp}{Term Premia decomposition}
  }
}
\usage{
data("NumOutEx")
}
\description{
Numerical outputs for JPS-based model using Brazilian data
}
\keyword{-}
\keyword{Brazil}
\keyword{JPS}
\keyword{Numerical}
\keyword{for}
\keyword{model}
\keyword{outputs}
