% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputsBootstrap.R
\name{NumOutputs_Bootstrap}
\alias{NumOutputs_Bootstrap}
\title{Numerical outputs (IRFs, GIRFs, FEVD, and GFEVD) for bootstrap}
\usage{
NumOutputs_Bootstrap(
  ModelType,
  ModelParaBoot,
  InputsForOutputs,
  FactorLabels,
  Economies
)
}
\arguments{
\item{ModelType}{A character vector indicating the model type to be estimated.}

\item{ModelParaBoot}{A list of model parameter estimates (see the "Optimization" function) after a bootstrap draw}

\item{InputsForOutputs}{A list containing the necessary inputs for generating IRFs, GIRFs, FEVDs, GFEVDs and Term Premia.}

\item{FactorLabels}{A list of character vectors with labels for all variables in the model.}

\item{Economies}{A character vector containing the names of the economies included in the system.}
}
\description{
Numerical outputs (IRFs, GIRFs, FEVD, and GFEVD) for bootstrap
}
\keyword{internal}
