% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{OOS_Forecast}
\alias{OOS_Forecast}
\title{Perform out-of-sample forecast of bond yields}
\usage{
OOS_Forecast(
  ForHoriz,
  t_Last,
  ModelParaList,
  FactorLabels,
  Yields_FullSample,
  Economies,
  ModelType
)
}
\arguments{
\item{ForHoriz}{Forecast horizon-ahead (scalar)}

\item{t_Last}{Index of the last set of observations in the information set at a given forecasting round}

\item{ModelParaList}{List of model parameter estimates}

\item{FactorLabels}{A string-list based which contains all the labels of all the variables present in the model}

\item{Yields_FullSample}{Time-series of bond yields, complete set (J x T or CJ x T)}

\item{Economies}{String-vector containing the names of the economies which are part of the economic system}

\item{ModelType}{A string-vector containing the label of the model to be estimated}
}
\description{
Perform out-of-sample forecast of bond yields
}
\keyword{internal}
