% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{Optimization_Time}
\alias{Optimization_Time}
\title{Compute the time elapsed in the numerical optimization}
\usage{
Optimization_Time(start_time)
}
\arguments{
\item{start_time}{Starting time}
}
\description{
Compute the time elapsed in the numerical optimization
}
\keyword{internal}
