% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{OrthoReg_JLL}
\alias{OrthoReg_JLL}
\title{Get coefficients from the orthogonalized regressions}
\usage{
OrthoReg_JLL(JLLinputs, FacSet, FactorLab_NonOrth, FactorLab_JLL)
}
\arguments{
\item{JLLinputs}{List of necessary inputs to estimate JLL-based setups}

\item{FacSet}{Set of factors used in the estimation of JLL-based setups}

\item{FactorLab_NonOrth}{Variable labels of the non-orthogonalized risk factors}

\item{FactorLab_JLL}{Variable labels of the orthogonalized risk factors}
}
\description{
Get coefficients from the orthogonalized regressions
}
\keyword{internal}
