% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLL.R
\name{OrthoVAR_JLL}
\alias{OrthoVAR_JLL}
\title{VAR(1) with orthogonalized factors (JLL models)}
\usage{
OrthoVAR_JLL(NonOrthoFactors, JLLinputs, Ortho_Set, FactLabels, N)
}
\arguments{
\item{NonOrthoFactors}{Risk factors before the orthogonalization (FxT)}

\item{JLLinputs}{List of necessary inputs to estimate JLL-based setups}

\item{Ortho_Set}{Set of orthogonalized risk factors}

\item{FactLabels}{Variable labels of the orthogonalized risk factors}

\item{N}{number of country-specific spanned factors (scalar)}
}
\description{
VAR(1) with orthogonalized factors (JLL models)
}
\keyword{internal}
