% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{OutputConstruction}
\alias{OutputConstruction}
\title{Numerical outputs (variance explained, model fit, IRFs, GIRFs, FEVDs, GFEVDs, and risk premia decomposition)
for all models}
\usage{
OutputConstruction(
  ModelType,
  ModelPara,
  InputsForOutputs,
  FactorLabels,
  Economies,
  verbose = TRUE
)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelPara}{list of model parameter estimates (See the "Optimization" function)}

\item{InputsForOutputs}{list containing the desired horizon of analysis for the model fit, IRFs, GIRFs, FEVDs, GFEVDs,
and risk premia decomposition}

\item{FactorLabels}{string-list based which contains all the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{verbose}{Logical flag controlling function messaging. Default is TRUE.}
}
\description{
Numerical outputs (variance explained, model fit, IRFs, GIRFs, FEVDs, GFEVDs, and risk premia decomposition)
for all models
}
\keyword{internal}
