% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{OutputConstructionJoint}
\alias{OutputConstructionJoint}
\title{Numerical outputs (variance explained, model fit, IRFs, GIRFs, FEVDs, GFEVDs and risk premia decomposition)
for "joint Q" models}
\usage{
OutputConstructionJoint(
  ModelType,
  ModelPara,
  InputsForOutputs,
  FactorLabels,
  Economies
)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelPara}{list of model parameter estimates (see the "Optimization" function)}

\item{InputsForOutputs}{list conataining the desired horizon of analysis for the model fit, IRFs, GIRFs, FEVDs, GFEVDs
and risk premia decomposition}

\item{FactorLabels}{string-list based which contains all the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Numerical outputs (variance explained, model fit, IRFs, GIRFs, FEVDs, GFEVDs and risk premia decomposition)
for "joint Q" models
}
