% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputsBootstrap.R
\name{OutputConstructionJoint_BS}
\alias{OutputConstructionJoint_BS}
\title{Gathers all the model numerical ouputs after bootstrap for "joint Q" models}
\usage{
OutputConstructionJoint_BS(
  ModelType,
  ModelParaBoot,
  InputsForOutputs,
  FactorLabels,
  Economies
)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelParaBoot}{list of model parameter estimates (see the "Optimization" function) after a bootstrap draw}

\item{InputsForOutputs}{list conataining the desired inputs for the cunstruction of IRFs, GIRFs, FEVDs, and GFEVDs}

\item{FactorLabels}{string-list based which contains all the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Gathers all the model numerical ouputs after bootstrap for "joint Q" models
}
\keyword{internal}
