% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{Outputs2exportMLE}
\alias{Outputs2exportMLE}
\title{Prepares inputs to export}
\usage{
Outputs2exportMLE(
  Label_Multi_Models,
  Economies,
  RiskFactors,
  Yields,
  mat,
  ModelInputsGen,
  ModelInputsSpec,
  PdynPara,
  ModelType
)
}
\arguments{
\item{Label_Multi_Models}{string-vector containing the names of the multicountry setups}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{RiskFactors}{time series of risk factors (F x T). Could be stored in a list depending on the model}

\item{Yields}{matrix (CJ x T) or a list containing those matrices, where C is the number of countries,
J - the number of maturities and T - time series length \cr}

\item{mat}{vector of maturities (in years) used in the estimation}

\item{ModelInputsGen}{List of generic inputs}

\item{ModelInputsSpec}{List of specific inputs}

\item{PdynPara}{Model parameters estimated in the P-dynamics the}

\item{ModelType}{string-vector containing the label of the model to be estimated}
}
\description{
Prepares inputs to export
}
\keyword{internal}
