% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{ParaATSM_opt_ALL}
\alias{ParaATSM_opt_ALL}
\title{Update the list of parameters}
\usage{
ParaATSM_opt_ALL(Update_Temp, FF_opt, AuxVecSet_opt, ParaLabels)
}
\arguments{
\item{Update_Temp}{List of model parameter features updated}

\item{FF_opt}{llk after optimization}

\item{AuxVecSet_opt}{List containing features for estimation of several parameters after optimization}

\item{ParaLabels}{Several variable labels}
}
\description{
Update the list of parameters
}
\keyword{internal}
