% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableLabelsOptimization.R
\name{ParaLabels}
\alias{ParaLabels}
\title{Create the variable labels used in the estimation}
\usage{
ParaLabels(ModelType, WishStationarityQ)
}
\arguments{
\item{ModelType}{a string-vector containing the label of the model to be estimated}

\item{WishStationarityQ}{User must set "1" is she whises to impose the largest eigenvalue under the Q to be strictly
smaller than 1. Otherwise set "0"}
}
\description{
Create the variable labels used in the estimation
}
