% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_Para.R
\docType{data}
\name{ParaSetEx}
\alias{ParaSetEx}
\title{Example of parameter set after optimization}
\format{
list of inputs and outputs
\describe{
   \item{Inupts}{general model inputs}
   \item{ModEst}{model parameters estimates (JPS form)}
  }
}
\usage{
data("ParaSetEx")
}
\description{
JPS-based model for Brazilian data
}
\keyword{Brazilian}
\keyword{Illustrative}
\keyword{JPS}
\keyword{data}
\keyword{for}
\keyword{model}
