% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliarBootstrapFunctions.R
\name{PdynamicsSet_BS}
\alias{PdynamicsSet_BS}
\title{Compute some key parameters from the P-dynamics (Bootstrap set)}
\usage{
PdynamicsSet_BS(
  ModelType,
  AllFactorsUnderP,
  FactorLabels,
  Economies,
  JLLinputs = NULL,
  GVARinputs = NULL
)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{AllFactorsUnderP}{complete set of factors that may be used in the estimation of P (KxT)}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model (see "LabFac" function)}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{JLLinputs}{List containing the necessary inputs for the estimation of the JLL-based models (see "JLL" function). Default is set to NULL.}

\item{GVARinputs}{List containing the necessary inputs for the estimation of the GVAR-based models (see "GVAR" function). Default is set to NULL.}
}
\description{
Compute some key parameters from the P-dynamics (Bootstrap set)
}
