% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{PdynResid_BS}
\alias{PdynResid_BS}
\title{Compute some key parameters from the P-dynamics (Bootstrap set)}
\usage{
PdynResid_BS(ModelType, Economies, ModelPara_PE)
}
\arguments{
\item{ModelType}{A character vector containing the label of the model to be estimated}

\item{Economies}{A character vector containing the names of the economies included in the system.}

\item{ModelPara_PE}{Point estimate from the model parameters}
}
\description{
Compute some key parameters from the P-dynamics (Bootstrap set)
}
\keyword{internal}
