% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Methods_All.R
\name{plot.ATSMModelForecast}
\alias{plot.ATSMModelForecast}
\title{Plot method for ATSMModelForecast objects}
\usage{
\method{plot}{ATSMModelForecast}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ATSMModelForecast}}

\item{...}{Additional arguments (not used)}
}
\description{
Plot method for ATSMModelForecast objects
}
