% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinorOptimizationFunctions.R
\name{pos2x}
\alias{pos2x}
\title{Transform a positive number y to back to x by:}
\usage{
pos2x(y)
}
\arguments{
\item{y}{scalar}
}
\description{
Transform a positive number y to back to x by:
}
\references{
This function is based on the "pos2x" function by Le and Singleton (2018).\cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
 Available at: https://cepr.org/40029
}
