% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{Reg_demean}
\alias{Reg_demean}
\title{Perform a linear regression using demeaned variables}
\usage{
Reg_demean(LHS, RHS)
}
\arguments{
\item{LHS}{Left-hand side variables}

\item{RHS}{Right-hand side variables}
}
\value{
Beta coefficients
}
\description{
Perform a linear regression using demeaned variables
}
\keyword{internal}
