% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabasePrep.R
\name{RemoveNA}
\alias{RemoveNA}
\title{Exclude series that contain NAs}
\usage{
RemoveNA(YieldsData, MacroData, Economies)
}
\arguments{
\item{YieldsData}{List of country-specific bond yields}

\item{MacroData}{List of country-specific and global economic variables}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\value{
return the time series data that were not initially composed by NAs.
}
\description{
Exclude series that contain NAs
}
