% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{residY_original}
\alias{residY_original}
\title{Compute the residuals from the observational equation}
\usage{
residY_original(ModelParaPE, BFull, ModelType, Economies)
}
\arguments{
\item{ModelParaPE}{List of point estimates of the model parameter}

\item{BFull}{Matrix B of loadings (CJ x F or J x F)}

\item{ModelType}{A character vector indicating the model type to be estimated}

\item{Economies}{String-vector containing the names of the economies which are part of the economic system}
}
\description{
Compute the residuals from the observational equation
}
\keyword{internal}
