% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{rhoParas}
\alias{rhoParas}
\title{Compute  risk-neutral intercept and slope}
\usage{
rhoParas(ModelPara, N, ModelType, Economies)
}
\arguments{
\item{ModelPara}{list of model parameter estimates}

\item{N}{number of country-specific spanned factors}

\item{ModelType}{desired model type}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Compute  risk-neutral intercept and slope
}
\keyword{internal}
