% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{richardson_diff}
\alias{richardson_diff}
\title{Richardson extrapolation}
\usage{
richardson_diff(f, x, i, h, fx0, nlevels = 5L)
}
\arguments{
\item{f}{function to differentiate}

\item{x}{evaluation point}

\item{i}{coordinate index}

\item{h}{base step size}

\item{fx0}{f(x), precomputed}

\item{nlevels}{depth of extrapolation. Default is 5}
}
\description{
Richardson extrapolation
}
\keyword{internal}
