% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_RiskFacFull.R
\docType{data}
\name{RiskFacFull}
\alias{RiskFacFull}
\title{Data: Full set of risk factors - Candelon and Moura (2024, JFEC)}
\format{
matrix containing the full risk factors: (i) global unspanned factors (global economic activity and global inflation); (ii) domestic unspanned factors (economic activity and inflation); and (iii) domestic spanned factors (level, slope, and curvature).
Economic system is formed by Brazil, China, Mexico and Uruguay. The data have monthly frequency and span the period from June/2004 to January/2020.
}
\source{
\describe{
\item{Global unspanned factor}{ See \code{data("GlobalMacro")} for a detailed data description.}
\item{Domestic unspanned factor}{See \code{data("DomMacro")} for a detailed data description.}
\item{Domestic spanned factor}{First three principal components of each country set of bond yields. See \code{data("Yields")} for a detailed data description.}
}
}
\usage{
data("RiskFacFull")
}
\description{
Full set of risk factors data used in Candelon and Moura (2024, JFEC)
}
\references{
Candelon, B. and Moura, R. (2024) "A Multicountry Model of the Term Structures of Interest Rates with a GVAR". (Journal of Financial Econometrics)
}
\keyword{Factors}
\keyword{Risk}
