% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Factors.R, R/CM_Factors_2023.R
\docType{data}
\name{RiskFactors}
\alias{RiskFactors}
\title{Data: Risk Factors - Candelon and Moura (2021)}
\format{
matrix containing the  risk factors of the models

matrix containing the  risk factors of the models
}
\usage{
data("CM_Factors")

data("CM_Factors_2023")
}
\description{
Risk factors data used in Candelon and Moura (2021)

Risk factors data used in Candelon and Moura (2023)
}
\references{
Candelon, B. and Moura, R. "A Multi-Country Model of the Term Structures of Interest Rates with a GVAR".

Candelon, B. and Moura, R. "Sovereign yield curves and the COVID-19 in emerging markets".
}
\keyword{Factors}
\keyword{Risk}
