% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskFactorsPrep.R
\name{RiskFactorsPrep}
\alias{RiskFactorsPrep}
\title{Builds the complete set of time series of the risk factors (spanned and unspanned)}
\usage{
RiskFactorsPrep(
  FactorSet,
  Economies,
  FactorLabels,
  Initial_Date,
  Final_Date,
  DataFrequency
)
}
\arguments{
\item{FactorSet}{Factor set list (see e.g. "CM_Factors_GVAR" data file)}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Initial_Date}{Sample starting date (yyyy-mm-dd)}

\item{Final_Date}{Sample last date (yyyy-mm-dd)}

\item{DataFrequency}{character-based vector: "Daily All Days", "Daily Business Days", "Weekly", "Monthly", "Quarterly", "Annually"}
}
\value{
Risk factors used in the estimation of the desired ATSM
}
\description{
Builds the complete set of time series of the risk factors (spanned and unspanned)
}
\keyword{internal}
