% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{RMSE}
\alias{RMSE}
\title{Compute the root mean square error for all models}
\usage{
RMSE(ForecastOutputs)
}
\arguments{
\item{ForecastOutputs}{List of country-specific forecasts}
}
\description{
Compute the root mean square error for all models
}
\keyword{internal}
