% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{RMSEjoint}
\alias{RMSEjoint}
\title{Compute the root mean square error ("joint Q" models)}
\usage{
RMSEjoint(ForecastOutputs)
}
\arguments{
\item{ForecastOutputs}{List of country-specific forecasts (see "ForecastYieldsjointQ" function)}
}
\description{
Compute the root mean square error ("joint Q" models)
}
