% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{RMSEsep}
\alias{RMSEsep}
\title{Compute the root mean square error ("sep Q" models)}
\usage{
RMSEsep(ForecastOutputs)
}
\arguments{
\item{ForecastOutputs}{List of country-specific forecasts (see "ForecastYieldsSepQ" function)}
}
\description{
Compute the root mean square error ("sep Q" models)
}
\keyword{internal}
