% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{RootEigen}
\alias{RootEigen}
\title{Compute the root of the eigenvalue of K1h}
\usage{
RootEigen(K1_h, horiz, N)
}
\arguments{
\item{K1_h}{K1_h squared matrix}

\item{horiz}{Minimum horizon observed}

\item{N}{number of country-specific spanned factors}
}
\description{
Compute the root of the eigenvalue of K1h
}
\keyword{internal}
