% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{Rotate_Lat_Obs}
\alias{Rotate_Lat_Obs}
\title{Rotate latent states to observed ones}
\usage{
Rotate_Lat_Obs(X_List, U1, U0)
}
\arguments{
\item{X_List}{list of risk-neutral parameters: (i) intercept (A_X); slope (B_X) and volatility matrix (SS_X)}

\item{U1}{rotation matrix (N x N)}

\item{U0}{rotation vector (N x 1)}
}
\description{
Rotate latent states to observed ones
}
\keyword{internal}
