% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{SA_algorithm}
\alias{SA_algorithm}
\title{Stochastic approximation algorithm}
\usage{
SA_algorithm(
  K1Z_NoBC,
  RiskFactors,
  BRWlist,
  GVARinputs,
  JLLinputs,
  FactorLabels,
  Economies,
  ModelType,
  verbose
)
}
\arguments{
\item{K1Z_NoBC}{feedback matrix before bias-correction}

\item{RiskFactors}{A numeric matrix (T x F) representing the time series of risk factors.}

\item{BRWlist}{A list containing the necessary inputs for the BRW model estimation}

\item{GVARinputs}{List. Inputs for GVAR model estimation.}

\item{JLLinputs}{List. Inputs for JLL model estimation.}

\item{FactorLabels}{A list of character vectors with labels for all variables in the model.}

\item{Economies}{A character vector containing the names of the economies included in the system.}

\item{ModelType}{A character vector indicating the model type to be estimated.}

\item{verbose}{verbose Logical flag controlling function messaging.}
}
\description{
Stochastic approximation algorithm
}
\keyword{internal}
