% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{scaling_from_jacobian}
\alias{scaling_from_jacobian}
\title{Scaling vector computation}
\usage{
scaling_from_jacobian(J)
}
\arguments{
\item{J}{jacobian approximation term}
}
\description{
Scaling vector computation
}
\keyword{internal}
