% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{shrink_FeedMat_BC}
\alias{shrink_FeedMat_BC}
\title{Shrinking the largest eigenvalue}
\usage{
shrink_FeedMat_BC(K1Z_BC, K1Z_NoBC, ev_restr)
}
\arguments{
\item{K1Z_BC}{VAR (1) bias-corrected feedback matrix from Bauer, Rudebusch and, Wu (2012)}

\item{K1Z_NoBC}{VAR (1) with no bias-corrected feedback matrix from the selected ATSM}

\item{ev_restr}{maximum eigenvalue desired in the feedback matrix after the adjustment}
}
\description{
Shrinking the largest eigenvalue
}
\keyword{internal}
