% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{shrink_Phi}
\alias{shrink_Phi}
\title{Killan's VAR stationarity adjustment}
\usage{
shrink_Phi(Phi_tilde, Phi_hat, ev_restr)
}
\arguments{
\item{Phi_tilde}{VAR (1) bias-corrected feedback matrix from Bauer, Rudebusch and, Wu (2012)}

\item{Phi_hat}{unrestricted VAR(1) feedback matrix}

\item{ev_restr}{maximum eigenvalue desired in the feedback matrix after the adjustment}
}
\value{
stationary VAR(1)
}
\description{
Killan's VAR stationarity adjustment
}
\references{
Bauer, Rudebusch and, Wu (2012). "Correcting Estimation Bias in Dynamic Term Structure Models". \cr
This function is an adapted version of the"shrink_Phi" Matlab function available at Cynthia Wu's website
(https://sites.google.com/view/jingcynthiawu/).
}
\keyword{internal}
