% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{SpannedFactorsSepQ}
\alias{SpannedFactorsSepQ}
\title{Gather all spanned factors ("sep Q" models)}
\usage{
SpannedFactorsSepQ(ModelType, ModelPara, Economies, t0Sample, tlastObserved)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelPara}{set of model parameters}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{t0Sample}{index for the initial sample date}

\item{tlastObserved}{index for the last observation of the information set}
}
\description{
Gather all spanned factors ("sep Q" models)
}
\keyword{internal}
