% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinorOptimizationFunctions.R
\name{sqrtm_robust}
\alias{sqrtm_robust}
\title{Compute the square root of a matrix}
\usage{
sqrtm_robust(m)
}
\arguments{
\item{m}{squared matrix (KxK)}
}
\value{
squared matrix x (KxK) such that x%*%x = m
}
\description{
Compute the square root of a matrix
}
\references{
#' This function is a modified version of the "sqrtm_robust" function by Le and Singleton (2018). \cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
 Available at: https://cepr.org/40029
}
\keyword{internal}
