% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{StarFactors}
\alias{StarFactors}
\title{Generates the star variables necessary for the GVAR estimation}
\usage{
StarFactors(RiskFactors, Economies, W)
}
\arguments{
\item{RiskFactors}{time series of the risk factors (F x T)}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{W}{GVAR transition matrix (C x C)}
}
\value{
List contaning the star factors of each country of the economic system
}
\description{
Generates the star variables necessary for the GVAR estimation
}
\keyword{internal}
