% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsForOpt.R
\name{summary.ATSMModelInputs}
\alias{summary.ATSMModelInputs}
\title{Summary method for ATSMModelInputs objects}
\usage{
\method{summary}{ATSMModelInputs}(Object, ...)
}
\arguments{
\item{Object}{An object of class 'ATSMModelInputs'}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for ATSMModelInputs objects
}
