% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Methods_All.R
\name{summary.ATSMModelOutputs}
\alias{summary.ATSMModelOutputs}
\title{Summary method for ATSMModelOutputs objects}
\usage{
\method{summary}{ATSMModelOutputs}(object, ...)
}
\arguments{
\item{object}{An object of class 'ATSMModelOutputs'}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for ATSMModelOutputs objects
}
