% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{TermPremia}
\alias{TermPremia}
\title{Compute the term premia}
\usage{
TermPremia(ModelPara, avexp, ModelType, Economies)
}
\arguments{
\item{ModelPara}{list of model parameter estimates}

\item{avexp}{list containing the country-specific pure expected component}

\item{ModelType}{desired model type}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Compute the term premia
}
\keyword{internal}
