% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{TermPremiaDecomp}
\alias{TermPremiaDecomp}
\title{Decomposition of yields into the average of expected future short-term interest rate and risk premia for all models}
\usage{
TermPremiaDecomp(
  ModelPara,
  FactorLabels,
  ModelType,
  InputsForOutputs,
  Economies
)
}
\arguments{
\item{ModelPara}{list of model parameter estimates (see the "Optimization" function)}

\item{FactorLabels}{string-list based which contains all the labels of all the variables present in the model}

\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{InputsForOutputs}{list containing the desired horizon of analysis for the model fit, IRFs, GIRFs, FEVDs, GFEVDs,
and risk premia decomposition}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Decomposition of yields into the average of expected future short-term interest rate and risk premia for all models
}
\keyword{internal}
