% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Trade_2023.R
\docType{data}
\name{Trade_Flows}
\alias{Trade_Flows}
\title{Data: Trade Flows - Candelon and Moura (2023)}
\format{
list containing the bilateral trade flows
}
\usage{
data("CM_Trade")
}
\description{
Trade Flows data used in Candelon and Moura (2023)
}
\references{
Candelon, B. and Moura, R. (2023) "Sovereign yield curves and the COVID-19 in emerging markets". (Economic Modelling)
}
\keyword{Flows}
\keyword{Trade}
