% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Trade_covid.R
\docType{data}
\name{TradeFlows_covid}
\alias{TradeFlows_covid}
\title{Data: Trade Flows - Candelon and Moura (2023, EM)}
\format{
A list. Contains bilateral trade flow data for Brazil, India, Mexico, and Russia,
covering the sample period from 1948 to 2020.
}
\source{
International Monetary Fund - International Trade in Goods (IMTS) <https://data.imf.org/en/datasets/IMF.STA:IMTS>
}
\usage{
data("TradeFlows_covid")
}
\description{
Trade Flows data used in Candelon and Moura (2023, EM)
}
\references{
Candelon, B. and Moura, R. (2023) "Sovereign yield curves and the COVID-19 in emerging markets". (Economic Modelling)
}
\keyword{Flows}
\keyword{Trade}
