% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionML_vec.R
\name{True_BlockDiag}
\alias{True_BlockDiag}
\title{Transformation of the block diagonal parameters (true form)}
\usage{
True_BlockDiag(ParaValue, Const_Type, FactorLabels, Economies, GVARinputs)
}
\arguments{
\item{ParaValue}{Constrained parameter}

\item{Const_Type}{Type of constraint}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{GVARinputs}{Inputs used in the estimation of the GVAR-based models}
}
\description{
Transformation of the block diagonal parameters (true form)
}
\keyword{internal}
