% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionf_vectorized.R
\name{True_BoundDiag}
\alias{True_BoundDiag}
\title{Transformation of the bounded parameters (True form)}
\usage{
True_BoundDiag(ParaValue, Const_Type, lb, ub)
}
\arguments{
\item{ParaValue}{Constrained parameter value}

\item{Const_Type}{Type of constraint}

\item{lb}{lower bound}

\item{ub}{upper bound}
}
\description{
Transformation of the bounded parameters (True form)
}
\keyword{internal}
