% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionML_vec.R
\name{True_JLLstruct}
\alias{True_JLLstruct}
\title{Transformation of the JLL-related parameters (true form)}
\usage{
True_JLLstruct(ParaValue, Const_Type, FactorLabels, Economies, JLLinputs)
}
\arguments{
\item{ParaValue}{Constrained parameter value}

\item{Const_Type}{Type of constraint}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{JLLinputs}{Inputs used in the estimation of the JLL-based models}
}
\description{
Transformation of the JLL-related parameters (true form)
}
\keyword{internal}
