% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionf_vectorized.R
\name{True_Jordan}
\alias{True_Jordan}
\title{Transformation of the Jordan-related parameters (True form)}
\usage{
True_Jordan(ParaValue, Const_Type, FactorLabels, Economies)
}
\arguments{
\item{ParaValue}{Constrained parameter}

\item{Const_Type}{Type of constraint}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Transformation of the Jordan-related parameters (True form)
}
\keyword{internal}
