% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionML_vec.R
\name{True_jordan_OneCountry}
\alias{True_jordan_OneCountry}
\title{True function for a single-country specification}
\usage{
True_jordan_OneCountry(ParaValue, stationary = FALSE)
}
\arguments{
\item{ParaValue}{Constrained parameter}

\item{stationary}{impose stationarity. Default is FALSE}
}
\description{
True function for a single-country specification
}
\references{
Le, A., & Singleton, K. J. (2018). Small Package of Matlab Routines for
Estimation of Some Term Structure Models. EABCN Training School.\cr
This function offers an independent R implementation that is informed
by the conceptual framework outlined in Le and Singleton (2018), but adapted to the
present modeling context.
}
\keyword{internal}
