% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionf_vectorized.R
\name{True_PSD}
\alias{True_PSD}
\title{Transformation of a PSD matrix (true form)}
\usage{
True_PSD(ParaValue, Const_Type)
}
\arguments{
\item{ParaValue}{Constrained parameter value}

\item{Const_Type}{Type of constraint}
}
\description{
Transformation of a PSD matrix (true form)
}
\keyword{internal}
