% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionML_vec.R
\name{Update_ParaList}
\alias{Update_ParaList}
\title{Update parameters in the optimization process}
\usage{
Update_ParaList(
  x,
  ModelType,
  FactorLabels,
  Economies,
  JLLinputs = NULL,
  GVARinputs = NULL,
  ListInputSet
)
}
\arguments{
\item{x}{vector containing all the vectorized auxiliary parameters}

\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{JLLinputs}{Set of necessary inputs used in the estimation of the JLL-based models}

\item{GVARinputs}{Set of necessary inputs used in the estimation of the GVAR-based models}

\item{ListInputSet}{variable inputs used in the optimization (see "Optimization" function)}
}
\description{
Update parameters in the optimization process
}
\keyword{internal}
