% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEdensity.R
\name{Update_SSZ_JLL}
\alias{Update_SSZ_JLL}
\title{Update the variance-covariance matrix from the "JLL joint Sigma" model. Necessary for optimization}
\usage{
Update_SSZ_JLL(SSZ, Z, N, JLLinputs)
}
\arguments{
\item{SSZ}{Variance-covariance matrix from JLL model}

\item{Z}{complete set of spanned and unspanned factors (F x T)}

\item{N}{number of country-specific spanned factors}

\item{JLLinputs}{List of inputs from JLL models}
}
\description{
Update the variance-covariance matrix from the "JLL joint Sigma" model. Necessary for optimization
}
\keyword{internal}
