% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{VarianceExplainedSep}
\alias{VarianceExplainedSep}
\title{Percentage explained by the spanned factors of the variations in the set of observed yields for "sep Q" models}
\usage{
VarianceExplainedSep(ModelType, ModelPara, FactorLabels, Economies)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelPara}{List of model parameter estimates (see the "Optimization" function)}

\item{FactorLabels}{string-list based which contains all the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}
}
\description{
Percentage explained by the spanned factors of the variations in the set of observed yields for "sep Q" models
}
