% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR.R
\name{VARX}
\alias{VARX}
\title{Estimate a VARX(1,1,1)}
\usage{
VARX(GVARinputs, Factors_GVAR, DomLabels, StarLabels, GlobalLabels, N)
}
\arguments{
\item{GVARinputs}{List of inputs for GVAR-based models}

\item{Factors_GVAR}{list containing the set of eisk factors used in the estimation of the VARX models}

\item{DomLabels}{string-based vector containing label of the domestic risk factors}

\item{StarLabels}{string-based vector containing label of the star domestic risk factors}

\item{GlobalLabels}{string-based vector containing label of the global risk factors}

\item{N}{number of country-specific spanned factors (scalar)}
}
\description{
Estimate a VARX(1,1,1)
}
\keyword{internal}
