% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{Wished_Graphs_FEVDandGFEVD}
\alias{Wished_Graphs_FEVDandGFEVD}
\title{Extract list of desired graph features (IRFs anc GIRFs)}
\usage{
Wished_Graphs_FEVDandGFEVD(InputsForOutputs, OutType, ModelType)
}
\arguments{
\item{InputsForOutputs}{List of inputs for outputs}

\item{OutType}{Output types "FEVD", "GFEVD" and "FEVD Ortho"}

\item{ModelType}{desired model type}
}
\description{
Extract list of desired graph features (IRFs anc GIRFs)
}
\keyword{internal}
