% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBoundsBootstrap.R
\name{WishGraphs_FEVDandGFEVD_Boot}
\alias{WishGraphs_FEVDandGFEVD_Boot}
\title{Extract graphs of interest (bootstrap version)}
\usage{
WishGraphs_FEVDandGFEVD_Boot(InputsForOutputs, ModelType)
}
\arguments{
\item{InputsForOutputs}{list containing the desired inputs for the construction of IRFs, GIRFs, FEVDs, and GFEVDs}

\item{ModelType}{desired model type}
}
\description{
Extract graphs of interest (bootstrap version)
}
\keyword{internal}
