% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{withDefault}
\alias{withDefault}
\title{Check default value}
\usage{
withDefault(value, default)
}
\arguments{
\item{value}{parameter to be checked}

\item{default}{default value}
}
\description{
Check default value
}
\keyword{internal}
