% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinorOptimizationFunctions.R
\name{x2bound}
\alias{x2bound}
\title{Transform x to a number bounded btw lb and ub by:}
\usage{
x2bound(x, lb, ub, nargout)
}
\arguments{
\item{x}{number to be transformed (scalar)}

\item{lb}{lower bound (scalar)}

\item{ub}{upper bound (scalar)}

\item{nargout}{"1" or "2" (scalar)}
}
\description{
Transform x to a number bounded btw lb and ub by:
}
\references{
This function is based on the "x2bound" function by Le and Singleton (2018). \cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling).
 Available at: https://cepr.org/40029
}
\keyword{internal}
