% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinorOptimizationFunctions.R
\name{x2pos}
\alias{x2pos}
\title{Transform x to a positive number by: y = log(e^x + 1)}
\usage{
x2pos(x, nargout)
}
\arguments{
\item{x}{scalar or vector}

\item{nargout}{1 or 2}
}
\description{
Transform x to a positive number by: y = log(e^x + 1)
}
\references{
This function is based on the "x2pos" function by Le and Singleton (2018). \cr
 "A Small Package of Matlab Routines for the Estimation of Some Term Structure Models." \cr
 (Euro Area Business Cycle Network Training School - Term Structure Modelling)
 Available at: https://cepr.org/40029
}
\keyword{internal}
