% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{Y_Fit}
\alias{Y_Fit}
\title{Model-implied yields (cross-section)}
\usage{
Y_Fit(ALoad, BLoad, Spa_TS, MatLength, TDim, YieldLab)
}
\arguments{
\item{ALoad}{A loadings}

\item{BLoad}{B loadings}

\item{Spa_TS}{time series of spanned factors}

\item{MatLength}{length of the vector of maturities}

\item{TDim}{Time-series dimension}

\item{YieldLab}{Label of yields}
}
\description{
Model-implied yields (cross-section)
}
\keyword{internal}
