% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{Y_ModImp}
\alias{Y_ModImp}
\title{Model-implied yields (P-dynamics)}
\usage{
Y_ModImp(ALoad, BLoad, K0Z, K1Z, PdynFact, MatLength, TDim, YieldLab)
}
\arguments{
\item{ALoad}{A loadings}

\item{BLoad}{B loadings}

\item{K0Z}{intercept from the P-dynamics}

\item{K1Z}{feedback matrix from the P-dynamics}

\item{PdynFact}{time series of the risk-factors spanned factors}

\item{MatLength}{length of the vector of maturities}

\item{TDim}{Time-series dimension}

\item{YieldLab}{Label of yields}
}
\description{
Model-implied yields (P-dynamics)
}
\keyword{internal}
