% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBoundsBootstrap.R
\name{YieldBounds_FEVDandGFEVD}
\alias{YieldBounds_FEVDandGFEVD}
\title{Compute the confidence bounds for the model bond yield-related outputs}
\usage{
YieldBounds_FEVDandGFEVD(
  ModelBootstrap,
  quants,
  ModelType,
  ndraws,
  Horiz,
  FacDim,
  YieDim,
  LabFEVD,
  Economies
)
}
\arguments{
\item{ModelBootstrap}{numerical output set from the bootstrap analysis}

\item{quants}{quantile of the confidence bounds}

\item{ModelType}{desired model type}

\item{ndraws}{number of draws}

\item{Horiz}{horizon of numerical outputs}

\item{FacDim}{dimension of the risk factor set}

\item{YieDim}{dimension of the bond yield set}

\item{LabFEVD}{vector containing the labels "FEVD" and "GFEVD"}

\item{Economies}{Economies that are part of the economic system}
}
\description{
Compute the confidence bounds for the model bond yield-related outputs
}
\keyword{internal}
