% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{YieldFor}
\alias{YieldFor}
\title{Compile the bond yield forecast for any model type}
\usage{
YieldFor(
  ModelParaList,
  ForHoriz,
  Economies,
  FactorLabels,
  ForLabels,
  ModelType
)
}
\arguments{
\item{ModelParaList}{List of model parameter estimates}

\item{ForHoriz}{Forecast horizon (scalar)}

\item{Economies}{String-vector containing the names of the economies which are part of the economic system}

\item{FactorLabels}{A string-list based which contains all the labels of all the variables present in the model}

\item{ForLabels}{Forecast labels (string-based vector)}

\item{ModelType}{A string-vector containing the label of the model to be estimated}
}
\description{
Compile the bond yield forecast for any model type
}
\keyword{internal}
