% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Yields_covid.R
\docType{data}
\name{Yields_covid}
\alias{Yields_covid}
\title{Data: Yields - Candelon and Moura (2023)}
\format{
A matrix containing Government bond yields with maturities of 12, 24, 36, 60, and 120 months for Brazil, India, Mexico, and Russia.
The data are at weekly frequency and cover the period from March 22, 2020, to September 26, 2021.
}
\source{
Simulated data constructed using Bloomberg bond yield series.
}
\usage{
data("Yields_covid")
}
\description{
Bond yield data used in Candelon and Moura (2023)
}
\references{
Candelon, B. and Moura, R. (2023) "Sovereign yield curves and the COVID-19 in emerging markets". (Economic Modelling)
}
\keyword{Yields}
