% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{YieldsFit}
\alias{YieldsFit}
\title{Computes two measures of model fit for bond yields (all models)}
\usage{
YieldsFit(ModelType, ModelPara, FactorLabels, Economies)
}
\arguments{
\item{ModelType}{a string-vector containing the label of the model to be estimated}

\item{ModelPara}{List of model parameter estimates (See the "Optimization" function)}

\item{FactorLabels}{a string-list based which contains the labels of all the variables present in the model}

\item{Economies}{a string-vector containing the names of the economies which are part of the economic system}
}
\description{
Computes two measures of model fit for bond yields (all models)
}
\details{
"Model-implied yields" is the measure of fit based exclusively on the risk-neutral parameters, whereas the
"Model-Fit" takes into account both the risk-neutral and the physical parameters.
}
\references{
See, for instance, Jotiskhatira, Le and Lundblad (2015). "Why do interest rates in different currencies co-move?" (Journal of Financial Economics)
}
\keyword{internal}
