% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumOutputs.R
\name{ComputeIRFs}
\alias{ComputeIRFs}
\title{Compute IRFs of all models}
\usage{
ComputeIRFs(
  SIGMA,
  K1Z,
  BLoad,
  FactorLabels,
  FacDim,
  MatLength,
  IRFhoriz,
  YieldsLabel,
  ModelType,
  Economy = NULL,
  PI = NULL,
  Mode = FALSE
)
}
\arguments{
\item{SIGMA}{Variance-covariance matrix}

\item{K1Z}{Loading As}

\item{BLoad}{Loading Bs}

\item{FactorLabels}{List containing the label of factors}

\item{FacDim}{Dimension of the P-dynamics}

\item{MatLength}{Length of the maturity vector}

\item{IRFhoriz}{Horizon of the analysis}

\item{YieldsLabel}{Label of bond yields}

\item{ModelType}{Desired model type}

\item{Economy}{specific economy under study}

\item{PI}{matrix PI for JLL-based models}

\item{Mode}{allows for the orthogonalized version in the case of JLL-based models}
}
\description{
Compute IRFs of all models
}
\keyword{internal}
