% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataForEstimation.R
\name{DataForEstimation}
\alias{DataForEstimation}
\title{Retrieve data from Excel and build the database used in the model estimation}
\usage{
DataForEstimation(
  t0,
  tF,
  Economies,
  N,
  FactorLabels,
  ModelType,
  DataFrequency,
  DataPathMacro = NULL,
  DataPathYields = NULL,
  Wgvar = NULL
)
}
\arguments{
\item{t0}{Sample starting date (yyyy-mm-dd)}

\item{tF}{Sample last date (yyyy-mm-dd)}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{N}{Number of country-specific spanned factor (scalar)}

\item{FactorLabels}{String-list based which contains the labels of all the variables present in the model}

\item{ModelType}{String-vector containing the label of the model to be estimated}

\item{DataFrequency}{Character-based-vector. Avaialable options are: "Daily All Days", "Daily Business Days", "Weekly", "Monthly", "Quarterly", "Annually"}

\item{DataPathMacro}{Path of the Excel file conating the macroeconomic data (if any). The default is linked to the excel file present in the package.}

\item{DataPathYields}{Path of the Excel file conating the yields data (if any). The default is linked to the excel file present in the package.}

\item{Wgvar}{GVAR transition matrix, if GVAR type model is chosen; default is set to NULL.}
}
\value{
A list containing the
\enumerate{
\item time series of the complete set of bond yields (matrix, JxT or CJxT);
\item time series of the complete set risk factors (matrix, KxT);
\item 'GVARFactors': list of all variables that are used in the estimation of the VARX \cr
               (see e.g. 'CM_Factors_GVAR' file). If the estimated model type is not GVAR-based, then returns NULL.
}
}
\description{
Retrieve data from Excel and build the database used in the model estimation
}
\examples{
DomVar <- c("Eco_Act", "Inflation")
GlobalVar <- c("GBC", "CPI_OECD")
t0 <- "2006-09-01"
tF <-  "2019-01-01"
Economies <- c("China", "Brazil", "Mexico", "Uruguay", "Russia")
N <- 2
ModelType <- "JPS"
FactorLabels <-  LabFac(N, DomVar, GlobalVar, Economies, ModelType)
DataFrequency <- "Monthly"


DataModel <- DataForEstimation(t0, tF, Economies, N, FactorLabels, ModelType, DataFrequency)

}
