% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_DomMacro.R
\docType{data}
\name{DomMacro}
\alias{DomMacro}
\title{Data: domestic risk factors - Candelon and Moura (2024, JFEC)}
\format{
A matrix of country-specific risk factors (inflation and economic activity growth) for Brazil, China, Mexico, and Uruguay.
The data have monthly frequency and span the period from June/2004 to January/2020.
}
\source{
\describe{
\item{Inflation}{year-over-year variation from Consumer price index published by the International Monetary Fund <https://data.imf.org/en/datasets/IMF.STA:CPI>}
\item{Eonomic activity}{GDP leading indicator published by the OECD <https://www.oecd.org/en/data/indicators/composite-leading-indicator-cli.html>}
}
}
\usage{
data("DomMacro")
}
\description{
Domestic risk factors data used in Candelon and Moura (2024, JFEC)
}
\references{
Candelon, B. and Moura, R. (2024) "A Multicountry Model of the Term Structures of Interest Rates with a GVAR". (Journal of Financial Econometrics)
}
\keyword{Domestic}
\keyword{factors}
\keyword{risk}
\keyword{unspanned}
