% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Factors_GVAR.R
\docType{data}
\name{FactorsGVAR}
\alias{FactorsGVAR}
\title{Data: Risk Factors for the GVAR - Candelon and Moura (2021)}
\format{
list containing the variables used in the GVAR models
}
\usage{
data("CM_Factors_GVAR")
}
\description{
Risk factors data used in the GVAR models - Candelon and Moura (2021)
}
\references{
Candelon, B. and Moura, R. "A Multi-Country Model of the Term Structures of Interest Rates with a GVAR".
}
\keyword{Factors}
\keyword{GVAR}
\keyword{Risk}
