% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Folder_Creation.R
\name{FolderCreationPoint}
\alias{FolderCreationPoint}
\title{Creates the folders and the path in which the graphical outputs are stored (point estimate version)}
\usage{
FolderCreationPoint(ModelType, Economies)
}
\arguments{
\item{ModelType}{A character vector indicating the model type to be estimated.}

\item{Economies}{A character vector containing the names of the economies included in the system.}
}
\description{
Creates the folders and the path in which the graphical outputs are stored (point estimate version)
}
\keyword{internal}
