% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastYields.R
\name{ForecastYields}
\alias{ForecastYields}
\title{Gather bond yields forecasts for all the model types}
\usage{
ForecastYields(
  ModelType,
  ModelPara,
  InputsForOutputs,
  FactorLabels,
  Economies,
  DataFrequency,
  JLLinputs,
  GVARinputs,
  BRWinputs
)
}
\arguments{
\item{ModelType}{a string-vector containing the label of the model to be estimated}

\item{ModelPara}{List of model parameter estimates (See the "Optimization" function)}

\item{InputsForOutputs}{list conataining the desired horizon of analysis for the IRFs, GIRFs, FEVDs, and GFEVDs}

\item{FactorLabels}{a string-list based which contains all the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{DataFrequency}{text: "Daily All Days", "Daily Business Days", "Weekly", "Monthly", "Quarterly", "Annually"}

\item{JLLinputs}{list of necessary inputs for the estimation of JLL-based models (see "JLL" function)}

\item{GVARinputs}{list of necessary inputs for the estimation of GVAR-based models (see "GVAR" function)}

\item{BRWinputs}{list of necessary inputs for performing the bias-corrected estimation (see "Bias_Correc_VAR" function)}
}
\value{
List containg the following elements
\enumerate{
\item Out-of-sample forecasts of bond yields per forecast horizon
\item Out-of-sample forecast errors of bond yields per forecast horizon
\item Root mean square errors per forecast horizon
}
}
\description{
Gather bond yields forecasts for all the model types
}
\examples{
# See examples in the vignette file of this package (Section 4).


}
