% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionML_vec.R
\name{GetTruePara}
\alias{GetTruePara}
\title{Map auxiliary (unconstrained) parameters a to constrained parameters b}
\usage{
GetTruePara(
  ParaValue,
  Const_Type_Full,
  FactorLabels,
  Economies,
  JLLinputs = NULL,
  GVARinputs = NULL
)
}
\arguments{
\item{ParaValue}{unconstrained auxiliary parameter}

\item{Const_Type_Full}{One of the following options:
\itemize{
\item 'Jordan'
\item 'Jordan; stationary'
\item 'Jordan MultiCountry'
\item 'Jordan MultiCountry; stationary'
\item 'psd';
\item 'BlockDiag'
\item 'JLLstructure'
}}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{JLLinputs}{Inputs used in the estimation of the JLL-based models}

\item{GVARinputs}{Inputs used in the estimation of the GVAR-based models}
}
\description{
Map auxiliary (unconstrained) parameters a to constrained parameters b
}
\keyword{internal}
