% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_GlobalMacro.R
\docType{data}
\name{GlobalMacro}
\alias{GlobalMacro}
\title{Data: Risk Factors - Candelon and Moura (2024, JFEC)}
\format{
A matrix containing the time series of global risk factors, namely global economic activity and global inflation.
The data have monthly frequency and span the period from June/2004 to January/2020.
}
\source{
\describe{
\item{Global economic activity}{Lutz Kilian’s index of global real economic activity <https://sites.google.com/site/lkilian2019/research/data-sets?authuser=0>}
\item{Global inflation}{year-over-year variation of the OECD aggregated CPI containing all its state members <https://www.oecd.org/en/data/indicators/inflation-cpi.html>}
}
}
\usage{
data("GlobalMacro")
}
\description{
Global risk factors data used in Candelon and Moura (2024, JFEC)
}
\references{
Candelon, B. and Moura, R. (2024) "A Multicountry Model of the Term Structures of Interest Rates with a GVAR". (Journal of Financial Econometrics)
}
\keyword{Global}
\keyword{factors}
\keyword{risk}
