% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Factors_GVAR_2023.R
\docType{data}
\name{GVARFactors}
\alias{GVARFactors}
\title{Data: Risk Factors for the GVAR - Candelon and Moura (2023)}
\format{
list containing the variables used in the GVAR models
}
\usage{
data("CM_Factors_GVAR_2023")
}
\description{
Risk factors data used in the GVAR models - Candelon and Moura (2023)
}
\references{
Candelon, B. and Moura, R. "Sovereign yield curves and the COVID-19 in emerging markets".
}
\keyword{Factors}
\keyword{GVAR}
\keyword{Risk}
