% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionf_vectorized.R
\name{ImposeStat_True}
\alias{ImposeStat_True}
\title{Makes sure that the stationary constraint under the risk-neutral measure is preserved}
\usage{
ImposeStat_True(x, K1Q)
}
\arguments{
\item{x}{parameter of interest (scalar or matrix)}

\item{K1Q}{risk-neutral feedback matrix}
}
\description{
Makes sure that the stationary constraint under the risk-neutral measure is preserved
}
\keyword{internal}
