% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{IRFandGIRFgraphs}
\alias{IRFandGIRFgraphs}
\title{IRFs graphs for all models}
\usage{
IRFandGIRFgraphs(
  ModelType,
  NumOut,
  WishPdynamicsgraphs,
  WishYieldsgraphs,
  IRFhoriz,
  PathsGraphs,
  OutputType,
  Economies
)
}
\arguments{
\item{ModelType}{a string-vector containing the label of the model to be estimated}

\item{NumOut}{list of computed outputs containing the model fit, IRFs, FEVDs, GIRFs, and GFEVDs}

\item{WishPdynamicsgraphs}{binary variable: set 1, if the user wishes graphs to be generated; or set 0, otherwise}

\item{WishYieldsgraphs}{binary variable: set 1, if the user wishes graphs to be generated; or set 0, otherwise}

\item{IRFhoriz}{single numerical vector containing the desired horizon of analysis for the IRFs}

\item{PathsGraphs}{Path of the folder in which the graphs will be saved}

\item{OutputType}{Available options are 'IRF' and 'GIRF'}

\item{Economies}{a string-vector containing the names of the economies which are part of the economic system}
}
\description{
IRFs graphs for all models
}
\keyword{internal}
