% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examples_Out.R
\docType{data}
\name{Out}
\alias{Out}
\title{Complete list of several outputs from an ATSM}
\format{
several model classes
\describe{
\item{ModelParaList}{List of parameter estimates of the selected ATSM}
   \item{ATSMinputs}{General inputs from an ATSM}
   \item{Forecasts}{List of forecast outputs}
   \item{NumOut}{List of numerical outputs}
   \item{Bootstrap}{List of set of outputs after bootstrap}
  }
}
\usage{
data("Out_Example")
}
\description{
Example for illustration used in the package vignette
}
\keyword{examples}
