% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{ParaLabelsOpt}
\alias{ParaLabelsOpt}
\title{Create the variable labels used in the estimation}
\usage{
ParaLabelsOpt(ModelType, WishStationarityQ, MLEinputs, BS_outputs = FALSE)
}
\arguments{
\item{ModelType}{a string-vector containing the label of the model to be estimated}

\item{WishStationarityQ}{User must set TRUE is she wishes to impose the largest eigenvalue under the Q to be strictly
smaller than 1. Otherwise set FALSE}

\item{MLEinputs}{Set of inputs that are necessary to the log-likelihood function}

\item{BS_outputs}{Generates simplified output list in the bootstrap setting. Default is set to FALSE.}
}
\description{
Create the variable labels used in the estimation
}
\keyword{internal}
