% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_weights_one_country.R
\name{pca_weights_one_country}
\alias{pca_weights_one_country}
\title{Weight matrix from principal components}
\usage{
pca_weights_one_country(Y, Economy)
}
\arguments{
\item{Y}{A numeric matrix (T x J) representing the time series of bond yields of a specific country}

\item{Economy}{String-vector containing the name of a single economy.}
}
\value{
A matrix (J x J) that corresponds to the eigenvectors of the variance-covariance matrix of yields
}
\description{
Weight matrix from principal components
}
\examples{
data("CM_Yields")
pca_weights_one_country(Yields, Economy= "Brazil")
}
