% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Methods_All.R
\name{print.ATSMModelInputs}
\alias{print.ATSMModelInputs}
\title{Print method for ATSMModelInputs objects}
\usage{
\method{print}{ATSMModelInputs}(x, ...)
}
\arguments{
\item{x}{An object of class 'ATSMModelInputs'}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for ATSMModelInputs objects
}
