% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Factors.R
\docType{data}
\name{RiskFactors}
\alias{RiskFactors}
\title{Data: Risk Factors - Candelon and Moura (forthcoming, JFEC)}
\format{
matrix containing the  risk factors of the models
}
\usage{
data("CM_Factors")
}
\description{
Risk factors data used in Candelon and Moura (forthcoming, JFEC)
}
\references{
Candelon, B. and Moura, R. (Forthcoming) "A Multicountry Model of the Term Structures of Interest Rates with a GVAR". (Journal of Financial Econometrics)
}
\keyword{Factors}
\keyword{Risk}
