% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Factors.R
\docType{data}
\name{RiskFactors}
\alias{RiskFactors}
\title{Data: Risk Factors - Candelon and Moura (2021)}
\format{
matrix containing the  risk factors of the models
}
\usage{
data("CM_Factors")
}
\description{
Risk factors data used in Candelon and Moura (2021)
}
\references{
Candelon, B. and Moura, R. "A Multi-Country Model of the Term Structures of Interest Rates with a GVAR".
}
\keyword{Factors}
\keyword{Risk}
