% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalOutputs.R
\name{RiskFactorsGraphs}
\alias{RiskFactorsGraphs}
\title{Spanned and unspanned factors plot}
\usage{
RiskFactorsGraphs(ModelType, ModelOutputs, Economies, FactorLabels)
}
\arguments{
\item{ModelType}{string-vector containing the label of the model to be estimated}

\item{ModelOutputs}{list of model parameter estimates (see the "Optimization" function)}

\item{Economies}{string-vector containing the names of the economies which are part of the economic system}

\item{FactorLabels}{string-list based which contains the labels of all the variables present in the model}
}
\description{
Spanned and unspanned factors plot
}
\keyword{internal}
