% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskFactorsPrep.R
\name{RiskFactorsPrep}
\alias{RiskFactorsPrep}
\title{Builds the complete set of time series of the risk factors (spanned and unspanned)}
\usage{
RiskFactorsPrep(
  FactorSet,
  Economies,
  FactorLabels,
  Initial_Date,
  Final_Date,
  DataFrequency
)
}
\arguments{
\item{FactorSet}{Factor set list (see e.g. "GVARFactors" data file)}

\item{Economies}{A character vector containing the names of the economies included in the system.}

\item{FactorLabels}{A list of character vectors with labels for all variables in the model.}

\item{Initial_Date}{Start date of the sample period in the format yyyy-mm-dd}

\item{Final_Date}{End date of the sample period in the format yyyy-mm-dd}

\item{DataFrequency}{A character vector specifying the data frequency. Available options: "Daily All Days", "Daily Business Days",
"Weekly", "Monthly", "Quarterly", "Annually".}
}
\value{
Risk factors used in the estimation of the desired ATSM
}
\description{
Builds the complete set of time series of the risk factors (spanned and unspanned)
}
\keyword{internal}
