% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Trade.R
\docType{data}
\name{TradeFlows}
\alias{TradeFlows}
\title{Data: Trade Flows - Candelon and Moura (forthcoming, JFEC)}
\format{
list containing the bilateral trade flows
}
\usage{
data("CM_Trade")
}
\description{
Trade Flows data used in Candelon and Moura (forthcoming, JFEC)
}
\references{
Candelon, B. and Moura, R. "A Multicountry Model of the Term Structures of Interest Rates with a GVAR".
}
\keyword{Flows}
\keyword{Trade}
