% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Trade.R
\docType{data}
\name{TradeFlows}
\alias{TradeFlows}
\title{Data: Trade Flows - Candelon and Moura (2024, JFEC)}
\format{
A list. Contains bilateral trade flow data for Brazil, China, Mexico, and Uruguay,
covering the sample period from 1948 to 2019.
}
\source{
International Monetary Fund - International Trade in Goods (IMTS) <https://data.imf.org/en/datasets/IMF.STA:IMTS>
}
\usage{
data("TradeFlows")
}
\description{
Trade Flows data used in Candelon and Moura (2024, JFEC)
}
\references{
Candelon, B. and Moura, R. (2024) "A Multicountry Model of the Term Structures of Interest Rates with a GVAR". (Journal of Financial Econometrics)
}
\keyword{Flows}
\keyword{Trade}
