% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Yields.R
\docType{data}
\name{Yields}
\alias{Yields}
\title{Data: Yields - Candelon and Moura (2021)}
\format{
matrix containing the Yields of the models
}
\usage{
data("CM_Yields")
}
\description{
Yields data used in Candelon and Moura (2021)
}
\references{
Candelon, B. and Moura, R. "A Multi-Country Model of the Term Structures of Interest Rates with a GVAR".
}
\keyword{Yields}
