% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Yields.R, R/CM_Yields_2023.R
\docType{data}
\name{Yields}
\alias{Yields}
\title{Data: Yields - Candelon and Moura (2024, JFEC)}
\format{
matrix containing the Yields of the models

matrix containing the Yields of the models
}
\usage{
data("CM_Yields")

data("CM_Yields_2023")
}
\description{
Yields data used in Candelon and Moura (2024, JFEC)

Bond yield data used in Candelon and Moura (2023)
}
\references{
Candelon, B. and Moura, R. (Forthcoming) "A Multicountry Model of the Term Structures of Interest Rates with a GVAR". (Journal of Financial Econometrics)

Candelon, B. and Moura, R. (2023) "Sovereign yield curves and the COVID-19 in emerging markets". (Economic Modelling)
}
\keyword{Yields}
